### 2023-05-30 #5501 @editorialbot
[PRE REVIEW]: PASCal Python: A Principal Axis Strain Calculator
**Submitting author:** @MJCliffe (<a href="http://orcid.org/0000-0002-0408-7647">Matthew Cliffe</a>)
**Repository:** https://github.com/MJCliffe/PASCal
**Branch with paper.md** (empty if default branch): joss
**Version:** v1.0.0
