### 2023-05-30 #5501 @Kevin-Mattheus-Moerman
@MJCliffe thanks for this submission. I am the EiC for this track and here to help process initial steps. Your project is on the small end of the spectrum in terms of what is acceptable for JOSS. Could you provide a bit more information as to how this project is used and relevant to scientific research? I can see it can be a useful tool to quickly check and visualize particular quantities, but I am unsure if the outputs have a direct application to research. E.g. does it only "quickly show and visualize" things? Or can the results directly be used (or exported) e.g. to be integrated in scientific publications? A quick of your online demo seems to show graphs and tables formatted on the web-page which cannot easily be exported or used further. But perhaps I am missing something there. Thanks.
