### 2023-05-30 #5501 @MJCliffe
@Kevin-Mattheus-Moerman thanks! The closed source version has been used very widely and cited more than 350 times and with the figures are directly used in papers, including in the table of contents figures: e.g. https://pubs.acs.org/doi/10.1021/acs.chemmater.2c03625

I've tried to make it so that the tables on the web page can be easily copy pasted as tsv files and figures exported, but other format suggestions are very welcome. I'm not a GitHub or JOSS expert so thanks for your help.

I will resolve the DOI issue: I didn't realise books could have DOIs!
