+++ paper.bib
@@ -82,6 +82,7 @@
   editor = {Giacovazzo, C.},
   year = {2011},
   edition = {3rd},
+  doi = {10.1093/acprof:oso/9780199573653.001.0001},
   publisher = {{International Union of Crystallography}}
 }
 % == BibTeX quality report for Giacovazzo2011:
