### 2023-06-13 #5556 @editorialbot
[REVIEW]: PASCal Python: A Principal Axis Strain Calculator
**Submitting author:** @MJCliffe (<a href="http://orcid.org/0000-0002-0408-7647">Matthew Cliffe</a>)
**Repository:** https://github.com/MJCliffe/PASCal
**Branch with paper.md** (empty if default branch): joss
**Version:** v2.1.0
