### 2023-06-16 #5556 @MarDiehl
A few remarks for improving the manuscript

- URL for the plotly citation is missing
- if https://pascal-notts.azurewebsites.net/ is stable, it would be good to mention it in the paper (I don't think that JOSS has a policy on online tools, might make sense to agree on requirements)
- no unit tests to ensure that the calculations are correct.
