### 2023-06-27 #5556 @MJCliffe
Thanks @MarDiehl & @phibeck. I'll get started on these issues. My inline comments on these issues are below:

- URL for the plotly citation is missing

Will resolve.

- if https://pascal-notts.azurewebsites.net/ is stable, it would be good to mention it in the paper (I don't think that JOSS has a policy on online tools, might make sense to agree on requirements)

Cannot guarantee stability, which is why I put the Github link as first preference, but can also include the current host. Would be good to know the requirements (or policy if it exists).

- no unit tests to ensure that the calculations are correct.

Will implement.

- Functionality documentation because PASCal Python is an online tool. For normal software, I would consider the functional documentation incomplete (no API documentation), but the GUI is certainly usable.

Great - I can definitely include more complete documentation of the functionality. If you could be a bit more detailed as to what would be appropriate (e.g. a high quality example) I will do so.
