### 2023-06-27 #5556 @phibeck
Hi @MarDiehl, thanks for pointing that out, and @MJCliffe, for getting started on the comments.

- Re: functionality documentation, I think a good start would be to use docstrings that also specify input/out parameters. This would allow to create automated documentation using appropriate tools e.g. Sphinx if necessary.
- Re: policy on including the current link, I'll inquire.

@MJCliffe, since PASCal already has a publication from 2012 (https://doi.org/10.1107/S0021889812043026), I would like to ask you to provide a summary of all the changes with respect to this version, highlighting the main new developments.

Ping @lbluque, @hungpham2017 and @SunilAnandatheertha. How is your progress on the review? @hungpham2017 please create your checklist.
