### 2023-06-28 #5556 @MJCliffe
Hi @phibeck,

A summary of changes over the previous software as requested:
- Entirely rewritten codebase (from Fortran/Perl/Gnuplot to Python) which is now properly open source
- Features for fitting electrochemical data (i.e. as a function of charge), including appropriate units and equations of state
- Improved treatment of eigenvectors. All eigenvectors are now output rather than just a median value, a new algorithm matching eigenvectors from different pressure points has been included (beyond simple rank sorting)
- Improvements made to fitting, including bug fixes, more sophisticated starting position choice

The discussion about functionality documentation has prompted me to have a broader rethink. To improve the reusability of the code it will make sense to convert the web app into a more conventional python package. This will not alter the function of the web app, but will allow easier offline use and hopefully facilitate integration into other materials & crystallography software. I have spoken with a collaborator (@ml-evs) who has agreed to help make this change, and would become a contributing author. This will take a few weeks. Would this be acceptable?
