### 2023-06-30 #5556 @phibeck
Hi @MJCliffe,

thank you very much for the summary. I think the changes you have in mind for the package sound reasonable. Even though implementing them might only take a few weeks, I will pause the Review process for now (using the "paused" label) to prevent reviewers from potentially reviewing twice if the changes end up being substantial.

@MarDiehl, @lbluque, @hungpham2017 and @SunilAnandatheertha, please take note of the pausing of the review process. Please let me know if you remain willing to review this submission in a few weeks, and I will notify you once the review process continues.

@MJCliffe and @MarDiehl, regarding potentially unstable links in the manuscript: it is okay to include them using language like "At the time of writing the web app is hosted here...", but do mention stable links such as the repository or documentation first, or instructions how to host a user's own instance.
