### 2023-07-05 6709e92
Move files into package directory and add `pyproject.toml`
