+++ .github/workflows/ci.yml
@@ -30,6 +30,7 @@ jobs:
         python -m pip install --upgrade pip
         pip install -U setuptools wheel
         pip install -r requirements.txt -r requirements-test.txt
+        pip install -e .
 
     - name: Run tests
       run: pytest -rs -vvv
