### 2023-07-05 b637233
Move app, templates and static files into src directory
