+++ requirements.txt
@@ -1,25 +1,0 @@
-Bottleneck
-certifi
-click
-Flask
-itsdangerous
-Jinja2
-json5
-MarkupSafe
-numexpr
-numpy
-packaging
-pandas
-patsy
-pip
-plotly
-pyparsing
-python-dateutil
-pytz
-scipy
-setuptools
-six
-statsmodels
-tenacity
-Werkzeug
-wheel
