+++ .github/workflows/ci.yml
@@ -29,8 +29,7 @@ jobs:
       run: |
         python -m pip install --upgrade pip
         pip install -U setuptools wheel
-        pip install -r requirements.txt -r requirements-test.txt
-        pip install -e .
+        pip install -e .\[test\]
 
     - name: Run tests
       run: pytest -rs -vvv
