### 2023-07-05 d6a8d4e
Tweak README installation instructions
