### 2023-07-05 8540418
Remove version string for now
