+++ pyproject.toml
@@ -2,31 +2,33 @@
 requires = ["setuptools >= 42", "wheel"]
 build-backend = "setuptools.build_meta"
 
-[project]
+[tool.poetry]
 name = "PASCal"
+version = "2.0.0"
 description = "Principal Axis Strain Calculator (PASCal) is a web tool designed to help scientists analyse non-ambient lattice parameter data."
 readme = "README.md"
 keywords = ["materials", "chemistry", "fitting", "strain", "lattice"]
-license = {text = "MIT"}
-authors = [{ name = "Matthew Cliffe" }]
-dynamic = ["version"]
+license = "MIT"
+authors = [ "Matthew Cliffe <matthew.cliffe@nottingham.ac.uk>", ]
 classifiers = [
     "Programming Language :: Python :: 3",
     "Programming Language :: Python :: 3.10",
     "Intended Audience :: Science/Research",
     "Topic :: Scientific/Engineering",
 ]
-requires-python = '>=3.8'
+homepage = "https://pascal-notts.azurewebsites.net"
+repository = "https://github.com/MJCliffe/PASCal"
 
-dependencies = [
-    "Flask",
-    "Jinja2",
-    "numpy",
-    "pandas",
-    "plotly",
-    "scipy",
-    "statsmodels",
-]
+[tool.poetry.dependencies]
+python = "^3.8"
+Flask = "*"
+Jinja2 = "*"
+numpy = "*"
+pandas = "*"
+plotly = "*"
+scipy = "*"
+statsmodels = "*"
+    
 
 [project.optional-dependencies]
 test = [
@@ -41,5 +43,4 @@ dev = [
 ]
 
 [project.urls]
-"Homepage" = "https://pascal-notts.azurewebsites.net"
-"Repository" = "https://github.com/MJCliffe/PASCal"
+
