+++ .github/workflows/ci.yml
@@ -28,8 +28,7 @@ jobs:
     - name: Install dependencies
       run: |
         python -m pip install --upgrade pip
-        pip install -U setuptools wheel
-        pip install -e .\[test\]
+        poetry install --with tests
 
     - name: Run tests
       run: pytest -rs -vvv
