### 2023-07-07 c8945dd
Actually install poetry in CI
