+++ .github/workflows/ci.yml
@@ -28,6 +28,7 @@ jobs:
     - name: Install dependencies
       run: |
         python -m pip install --upgrade pip
+        pip install poetry
         poetry install --all-extras
 
     - name: Run tests
