### 2023-07-07 f868b8c
Explicitly add package to pyproject to allow src layout
