### 2023-07-07 b69d198
Explicitly run pytest through poetry
