+++ .github/workflows/ci.yml
@@ -32,4 +32,8 @@ jobs:
         poetry install --all-extras
 
     - name: Run tests
-      run: poetry run 'pytest -rs -vvv'
+      run: |
+        poetry run 'which python'
+        poetry run 'pip freeze'
+        pip freeze
+        poetry run 'pytest -rs -vvv'
