### 2023-08-16 c83422d
Refactor top-level functions into a utils module and begin adding tests and docs
