+++ src/PASCal/utils.py
@@ -1,0 +1,273 @@
+import numpy as np
+from typing import Union, Callable
+from functools import partial
+
+
+def round_array(var: np.ndarray, dec: int) -> Union[np.ndarray, float]:
+    """Rounding the number to desired decimal places
+    `round()` is more accurate at rounding float numbers than `np.round()`.
+
+    Parameters:
+        var: The array or scalar to round.
+        dec: The number of decimal places to round to.
+
+    Returns:
+        The rounded array or scalar.
+
+    """
+    if var.ndim == 0:
+        return round(var, dec)  # type: ignore
+
+    for indices, val in np.ndenumerate(var):
+        var[indices] = round(val, dec)
+    return var
+
+
+def orthomat(latt: np.ndarray) -> np.ndarray:
+    """Compute the corresponding change-of-basis transformation
+    (square matrix M) in E = M x A.
+
+    Parameters:
+        latt: The lattice parameters (a, b, c, α, β, γ) in Å and degrees, respectively.
+
+    Returns:
+        The change-of-basis matrix M.
+
+    """
+    orth = np.zeros((np.shape(latt)[0], 3, 3))
+    a, b, c = latt[:3]
+    alpha, beta, gamma = np.radians(latt[3:])
+    gaS = np.arccos(
+        (np.cos(alpha) * np.cos(beta) - np.cos(gamma)) / (np.sin(alpha) * np.sin(beta))
+    )
+    orth[0, 0] = 1 / (a * np.sin(beta) * np.sin(gaS))
+    orth[1, 0] = np.cos(gaS) / (b * np.sin(alpha) * np.sin(gaS))
+    orth[2, 0] = (
+        np.cos(alpha) * np.cos(gaS) / np.sin(alpha) + np.cos(beta) / np.sin(beta)
+    ) / (-1 * b * np.sin(gaS))
+    orth[1, 1] = 1 / (b * np.sin(alpha))
+    orth[2, 1] = -1 * np.cos(alpha) / (c * np.sin(alpha))
+    orth[2, 2] = 1 / c
+
+    return orth
+
+
+def cell_vol(latt: np.ndarray) -> float:
+    """Calculates the unit-cell volume.
+
+    Parameters:
+        latt: The lattice parameters (a, b, c, α, β, γ) in Å and degrees, respectively.
+
+    Returns:
+        The unit cell volume in Å³.
+    """
+    a, b, c = latt[:3]
+    alpha, beta, gamma = np.radians(latt[3:])
+    return (
+        a * b * c
+        * (1 - np.cos(alpha) ** 2 - np.cos(beta) ** 2 - np.cos(gamma) ** 2
+        + (2 * np.cos(alpha) * np.cos(beta) * np.cos(gamma)))** (0.5)
+    )
+
+
+def EmpEq(TP: np.ndarray, Epsilon0: np.ndarray, lambdaP: float, Pc: float, Nu: float) -> np.ndarray:
+    """Empirical fit for pressure input data.
+
+    Parameters:
+        TP: array of pressure data points,
+        Epsilon0: strain at critical pressure
+        lambdaP: compressibility in (GPa^-nu)
+        Pc: critical pressure (GPa)
+        Nu: rate of stiffening (0.5)
+
+    Returns:
+        The strain at each pressure point.
+
+    """
+    return Epsilon0 + (lambdaP * ((TP - Pc) ** Nu))
+
+
+def Comp(TP: np.ndarray, lambdaP: float, Pc: float, Nu: float) -> np.ndarray:
+    """Calculate the compressibility from the derivative -de/dp.
+
[…]
