+++ src/PASCal/app.py
@@ -3,7 +3,7 @@ import plotly
 import plotly.graph_objs as go
 import plotly.subplots
 import statsmodels.api as sm
-import PASCal
+import PASCal.utils
 import numpy as np
 import json
 import os
@@ -102,7 +102,7 @@ def output():
     TPx = np.array(TPx)
     TPxError = np.array(TPxError)
     Latt = np.stack(Latt)
-    Vol = PASCal.CellVol(Latt)  # volumes in (Angstrom^3)
+    Vol = PASCal.utils.cell_vol(Latt)  # volumes in (Angstrom^3)
     ## Data error handling
     if len(TPx) < 2:
         raise Exception("Too few data points")
@@ -110,7 +110,7 @@ def output():
     if DataType == "Pressure":
         if len(TPx) < 4:
             warning.append(
-                "At least as many data points as parameters are needed for a fit to be carried out (e.g. 3 for 3rd order Birch-Murnaghan, 4 for empirical pressure fitting). As PASCal calculates errors from derivatives, more data points than parameters are needed for error estimates."
+                "At least as many data points as parameters are needed for a fit to be carried out (e.g. 3 for 3rd order Birch-Murnaghan, 4 for empirical pressure fitting). As PASCal.utils calculates errors from derivatives, more data points than parameters are needed for error estimates."
             )
         if UsePc == "True":
             if np.amin(TPx) < PcVal:
@@ -127,12 +127,12 @@ def output():
         if len(TPx) - 2 < DegPolyCap:
             DegPolyCap = len(TPx) - 2
             warning.append(
-                "The maximum degree of the Chebyshev strain polynomial has been lowered. At least as many data points as parameters are needed for a fit to be carried out. As PASCal calculates errors from derivatives, more data points than parameters are needed for error estimates."
+                "The maximum degree of the Chebyshev strain polynomial has been lowered. At least as many data points as parameters are needed for a fit to be carried out. As PASCal.utils calculates errors from derivatives, more data points than parameters are needed for error estimates."
             )
         if len(TPx) - 2 < DegPolyVol:
             DegPolyVol = len(TPx) - 2
             warning.append(
-                "The maximum degree of the Chebyshev volume polynomial has been lowered. At least as many data points as parameters are needed for a fit to be carried out. As PASCal calculates errors from derivatives, more data points than parameters are needed for error estimates."
+                "The maximum degree of the Chebyshev volume polynomial has been lowered. At least as many data points as parameters are needed for a fit to be carried out. As PASCal.utils calculates errors from derivatives, more data points than parameters are needed for error estimates."
             )
 
     percent = 1e2  # conversions
@@ -145,7 +145,7 @@ def output():
     ####
 
     ### Strain calculations
-    Orth = PASCal.Orthomat(Latt[:])  # cell in orthogonal axes
+    Orth = PASCal.utils.orthomat(Latt[:])  # cell in orthogonal axes
     if EulerianStrain:  # Strain calculation
[…]

[…]
