### 2023-08-17 ad0dd56
More WIP refactoring
