### 2023-08-17 7311847
Some sanity checks with the old code
