+++ tests/test_utils.py
@@ -1,4 +1,4 @@
-from PASCal.utils import round_array, orthomat, cell_vol
+from PASCal.utils import round_array, orthomat, cell_vol, Round, Orthomat, CellVol
 import numpy as np
 import random
 from pytest import fixture
@@ -20,6 +20,7 @@ def test_round_array():
         rounded = round(c, dec)
         for _I in [I1, I2, I3]:
             assert round_array(_I * c, 1).all() == (rounded * _I).all()
+            assert Round(_I * c, 1).all() == round_array(_I * c, 1).all()
 
 def test_orthomat():
     lattice = np.array([[1, 1, 1, 90, 90, 90]])
@@ -51,6 +52,7 @@ def test_orthomat():
     assert orth.all() == np.array([M]).all()
     orth = orthomat(np.hstack((lattice, lattice)))
     assert orth.all() == np.array([M, M]).all()
+    assert orth.all() == Orthomat(np.hstack((lattice, lattice))).all()
 
 def test_cell_vol():
     lattice = np.array([[1, 1, 1, 90, 90, 90]])
@@ -58,3 +60,4 @@ def test_cell_vol():
     assert cell_vol(lattice).all() == np.array([1]).all()
     assert cell_vol(lattice2).all() == np.array([8]).all()
     assert cell_vol(np.hstack((lattice, lattice2))).all() == np.array([1, 8]).all()
+    assert cell_vol(np.hstack((lattice, lattice2))).all() == CellVol(np.hstack((lattice, lattice2))).all()
