+++ src/PASCal/utils.py
@@ -268,3 +268,200 @@ def Indicatrix(PrinComp: np.ndarray):
     Y = R * np.sin(THETA) * np.sin(PHI)
     Z = R * np.cos(THETA)
     return maxIn, R, X, Y, Z
+
+def Round(var, dec):
+    # Rounding the number to desired decimal places
+    # round() is more accurate at rounding float numbers than np.round()
+    if var.ndim == 0:
+        var = round(var, dec)
+    if var.ndim == 1:
+        for i in range(var.shape[0]):
+            var[i] = round(var[i], dec)
+    if var.ndim == 2:
+        for i in range(var.shape[0]):
+            for j in range(var.shape[1]):
+                var[i][j] = round(var[i][j], dec)
+    if var.ndim == 3:
+        for i in range(var.shape[0]):
+            for j in range(var.shape[1]):
+                for k in range(var.shape[2]):
+                    var[i][j][k] = round(var[i][j][k], dec)
+    if var.ndim < 0:
+        raise TypeError(var, " has incorrect dimensions in rounding")
+    if var.ndim > 3:
+        raise TypeError(var, " has incorrect dimensions in rounding")
+    return var
+
+
+def Orthomat(Latt):
+    # Compute the corresponding change-of-basis transformation (square matrix M) in E = M x A
+    # Latt: a b c alpha beta gamma (Angstrom, degrees)
+    orth = np.zeros((np.shape(Latt)[0], 3, 3))
+    alpha = Latt[:, 3] * (np.pi / 180)
+    beta = Latt[:, 4] * (np.pi / 180)
+    gamma = Latt[:, 5] * (np.pi / 180)
+    gaS = np.arccos(
+        (np.cos(alpha) * np.cos(beta) - np.cos(gamma)) / (np.sin(alpha) * np.sin(beta))
+    )
+    orth[:, 0, 0] = 1 / (Latt[:, 0] * np.sin(beta) * np.sin(gaS))
+    orth[:, 1, 0] = np.cos(gaS) / (Latt[:, 1] * np.sin(alpha) * np.sin(gaS))
+    orth[:, 2, 0] = (
+        np.cos(alpha) * np.cos(gaS) / np.sin(alpha) + np.cos(beta) / np.sin(beta)
+    ) / (-1 * Latt[:, 2] * np.sin(gaS))
+    orth[:, 1, 1] = 1 / (Latt[:, 1] * np.sin(alpha))
+    orth[:, 2, 1] = -1 * np.cos(alpha) / (Latt[:, 2] * np.sin(alpha))
+    orth[:, 2, 2] = 1 / Latt[:, 2]
+
+    return orth
+
+
+def CellVol(LattPam):
+    # Calculate the unit-cell volume
+    # Lattice parameters a b c al be ga (Angstrom, degrees)
+    vol = np.zeros((LattPam.shape[0]))
+    for i in range(0, LattPam.shape[0]):
+        Latt = LattPam[i]
+        vol[i] = (
+            Latt[0]
+            * Latt[1]
+            * Latt[2]
+            * (
+                (1 - np.cos(Latt[3] * (np.pi / 180)) ** 2)
+                - (np.cos(Latt[4] * (np.pi / 180)) ** 2)
+                - (np.cos(Latt[5] * (np.pi / 180)) ** 2)
+                + (
+                    2
+                    * np.cos(Latt[3] * (np.pi / 180))
+                    * np.cos(Latt[4] * (np.pi / 180))
+                    * np.cos(Latt[5] * (np.pi / 180))
+                )
[…]
