+++ .github/workflows/ci.yml
@@ -48,7 +48,7 @@ jobs:
     permissions:
       contents: none
     runs-on: ubuntu-latest
-    needs: build
+    needs: tests
     environment:
       name: "Test deployment"
       url: ${{ steps.deploy-to-webapp.outputs.webapp-url }}
