### 2023-08-17 111fdeb
Remove startup command
