+++ .github/workflows/ci.yml
@@ -33,8 +33,6 @@ jobs:
 
     - name: Run tests
       run: |
-        which python
-        pip freeze
         poetry run pytest
 
     - name: Upload artifact for deployment
@@ -66,4 +64,3 @@ jobs:
         with:
           app-name: ${{ env.AZURE_WEBAPP_NAME }}
           publish-profile: ${{ secrets.AZURE_WEBAPP_PUBLISH_PROFILE }}
-          startup-command: poetry run flask run
