### 2023-08-17 bf84390
Use test publish profile initially
