### 2023-08-17 ee8fe3b
Trick Azure into donwloading requirements
