+++ .github/workflows/ci.yml
@@ -24,9 +24,10 @@ jobs:
         cache-dependency-path: |
           requirements*.txt
 
-
     - name: Install dependencies
       run: |
+        python -m venv venv
+        source venv/bin/activate
         python -m pip install --upgrade pip
         pip install poetry
         poetry install --all-extras
@@ -35,11 +36,17 @@ jobs:
       run: |
         poetry run pytest
 
+    - name: Export poetry requirements as txt
+      run: |
+        poetry export -f requirements.txt -o requirements.txt --without-hashes
+
     - name: Upload artifact for deployment
       uses: actions/upload-artifact@v3
       with:
         name: PASCal - test
-        path: .
+        path: |
+          .
+          !venv/
 
   deploy:
     if: github.repository == 'MJCliffe/PASCal'
