+++ .github/workflows/ci.yml
@@ -26,8 +26,6 @@ jobs:
 
     - name: Install dependencies
       run: |
-        python -m venv venv
-        source venv/bin/activate
         python -m pip install --upgrade pip
         pip install poetry
         poetry install --all-extras
