+++ requirements.txt
@@ -1,0 +1,22 @@
+blinker==1.6.2
+click==8.1.4
+colorama==0.4.6
+flask==2.3.2
+importlib-metadata==6.7.0
+itsdangerous==2.1.2
+jinja2==3.1.2
+markupsafe==2.1.3
+numpy==1.24.4
+packaging==23.1
+pandas==2.0.3
+patsy==0.5.3
+plotly==5.15.0
+python-dateutil==2.8.2
+pytz==2023.3
+scipy==1.9.3
+six==1.16.0
+statsmodels==0.14.0
+tenacity==8.2.2
+tzdata==2023.3
+werkzeug==2.3.6
+zipp==3.15.0
