### 2023-08-17 602c8aa
Add workflow_dispatch triggered build for live server
