+++ .github/workflows/deploy.yml
@@ -1,0 +1,64 @@
+name: Deploy to main server
+
+on:
+  workflow_dispatch:
+
+jobs:
+
+  build:
+
+    steps:
+    - uses: actions/checkout@v3
+      with:
+        submodules: true
+
+
+    - name: Set up Python 3.8
+      uses: actions/setup-python@v4
+      with:
+        python-version: '3.8'
+        cache: 'pip'
+        cache-dependency-path: |
+          requirements*.txt
+
+    - name: Install dependencies
+      run: |
+        python -m pip install --upgrade pip
+        pip install poetry
+        poetry install
+
+    - name: Export poetry requirements as txt
+      run: |
+        poetry export -f requirements.txt -o requirements.txt --without-hashes
+
+    - name: Upload artifact for deployment
+      uses: actions/upload-artifact@v3
+      with:
+        name: PASCal - main
+        path: |
+          .
+          !venv/
+
+  deploy:
+    if: github.repository == 'MJCliffe/PASCal'
+    permissions:
+      contents: none
+    runs-on: ubuntu-latest
+    needs: tests
+    environment:
+      name: "Main deployment"
+      url: ${{ steps.deploy-to-webapp.outputs.webapp-url }}
+
+    steps:
+      - name: Download artifact from build job
+        uses: actions/download-artifact@v3
+        with:
+          name: PASCal - main
+          path: .
+
+      - name: 'Deploy to Azure Web App'
+        id: deploy-to-webapp
+        uses: azure/webapps-deploy@v2
+        with:
+          app-name: ${{ env.AZURE_WEBAPP_NAME }}
+          publish-profile: ${{ secrets.AZURE_WEBAPP_PUBLISH_PROFILE }}
