### 2023-08-17 d61b6cd
Also copy static files
