### 2023-08-17 b68579f
Only deploy test on pushes and fix main deployment
