+++ .github/workflows/deploy.yml
@@ -27,9 +27,16 @@ jobs:
         pip install poetry
         poetry install
 
-    - name: Export poetry requirements as txt
+    - name: Lock dependencies and prepare deployment
       run: |
+        # Azure expects to find the app, requirements and any static files at the top-level
+        # so we need to do some tweaks here to massage the repo into that format
         poetry export -f requirements.txt -o requirements.txt --without-hashes
+        # add self-install to end of requirements file
+        sh -c 'echo "." >> requirements.txt'
+        # copy app file so it is available as app:app
+        mv src/PASCal/app.py src/PASCal/static src/PASCal/templates .
+        ls -al
 
     - name: Upload artifact for deployment
       uses: actions/upload-artifact@v3
