### 2023-08-17 710222f
Test Azure deployment workflow (#7)
* First attempt at workflow

* Second attempt at workflow

* Remove startup command

* Use test publish profile initially

* Trick Azure into donwloading requirements

* Forget the venv

* Try with raw requirements.txt

* Add workflow_dispatch triggered build for live server

* Try to trick Azure

* Try to trick Azure take 2

* Tweak

* Also copy static files

* Try dynamically generating requirements file

* Only deploy test on pushes and fix main deployment
