+++ .github/workflows/deploy.yml
@@ -1,19 +1,17 @@
-name: CI tests
+name: Deploy to main server
 
 on:
-  push:
-    branches:
-    - Live
+  workflow_dispatch:
 
 jobs:
 
-  tests:
-    runs-on: ubuntu-latest
+  build:
 
     steps:
     - uses: actions/checkout@v3
       with:
         submodules: true
+
 
     - name: Set up Python 3.8
       uses: actions/setup-python@v4
@@ -27,11 +25,7 @@ jobs:
       run: |
         python -m pip install --upgrade pip
         pip install poetry
-        poetry install --all-extras
-
-    - name: Run tests
-      run: |
-        poetry run pytest
+        poetry install
 
     - name: Lock dependencies and prepare deployment
       run: |
@@ -47,7 +41,7 @@ jobs:
     - name: Upload artifact for deployment
       uses: actions/upload-artifact@v3
       with:
-        name: PASCal - test
+        name: PASCal - main
         path: |
           .
           !venv/
@@ -59,14 +53,14 @@ jobs:
     runs-on: ubuntu-latest
     needs: tests
     environment:
-      name: "Test deployment"
+      name: "Main deployment"
       url: ${{ steps.deploy-to-webapp.outputs.webapp-url }}
 
     steps:
       - name: Download artifact from build job
         uses: actions/download-artifact@v3
         with:
-          name: PASCal - test
+          name: PASCal - main
           path: .
 
       - name: 'Deploy to Azure Web App'
@@ -74,4 +68,4 @@ jobs:
         uses: azure/webapps-deploy@v2
         with:
           app-name: ${{ env.AZURE_WEBAPP_NAME }}
-          publish-profile: ${{ secrets.AZURE_TEST_WEBAPP_PUBLISH_PROFILE }}
+          publish-profile: ${{ secrets.AZURE_WEBAPP_PUBLISH_PROFILE }}
