### 2023-08-18 5905804
Update deploy.yml
Added a runs-on Ubuntu command
