+++ .github/workflows/deploy.yml
@@ -8,6 +8,7 @@ jobs:
   build:
 
     steps:
+    - runs-on: 'ubuntu-latest'
     - uses: actions/checkout@v3
       with:
         submodules: true
