### 2023-08-18 b50c683
Update deploy.yml
Moved 'runs-on'
