+++ .github/workflows/deploy.yml
@@ -6,9 +6,8 @@ on:
 jobs:
 
   build:
-
-    steps:
     - runs-on: 'ubuntu-latest'
+    steps:    
     - uses: actions/checkout@v3
       with:
         submodules: true
