### 2023-08-18 080c74c
Update deploy.yml
syntax tweaks runs-on
