+++ .github/workflows/deploy.yml
@@ -6,7 +6,7 @@ on:
 jobs:
 
   build:
-    - runs-on: 'ubuntu-latest'
+    runs-on: ubuntu-latest
     steps:    
     - uses: actions/checkout@v3
       with:
