### 2023-08-23 e10a1ac
Reworked all utility functions and broke off legacy module, with new regression tests
