+++ src/PASCal/utils.py
@@ -1,0 +1,267 @@
+import numpy as np
+from typing import Union, Tuple
+
+
+def round_array(var: np.ndarray, dec: int) -> Union[np.ndarray, float]:
+    """Rounding the number to desired decimal places
+    `round()` is more accurate at rounding float numbers than `np.round()`.
+
+    Parameters:
+        var: The array or scalar to round.
+        dec: The number of decimal places to round to.
+
+    Returns:
+        The rounded array or scalar.
+
+    """
+    if var.ndim == 0:
+        return round(var, dec)  # type: ignore
+
+    for indices, val in np.ndenumerate(var):
+        var[indices] = round(val, dec)
+    return var
+
+
+def orthomat(lattices: np.ndarray) -> np.ndarray:
+    """Compute the corresponding change-of-basis transformations
+    (square matrix M) in E = M x A.
+
+    Parameters:
+        lattices: An Nx6 array of lattice parameters (a, b, c, α, β, γ)
+            in Å and degrees, respectively.
+
+    Returns:
+        An array of change-of-basis matrices M for the N cells.
+
+    """
+    orth = np.zeros((np.shape(lattices)[0], 3, 3))
+    alphas, betas, gammas = (
+        np.radians(lattices[:, 3]),
+        np.radians(lattices[:, 4]),
+        np.radians(lattices[:, 5]),
+    )
+    gaS = np.arccos(
+        (np.cos(alphas) * np.cos(betas) - np.cos(gammas))
+        / (np.sin(alphas) * np.sin(betas))
+    )
+    orth[:, 0, 0] = 1 / (lattices[:, 0] * np.sin(betas) * np.sin(gaS))
+    orth[:, 1, 0] = np.cos(gaS) / (lattices[:, 1] * np.sin(alphas) * np.sin(gaS))
+    orth[:, 2, 0] = (
+        np.cos(alphas) * np.cos(gaS) / np.sin(alphas) + np.cos(betas) / np.sin(betas)
+    ) / (-1 * lattices[:, 1] * np.sin(gaS))
+    orth[:, 1, 1] = 1 / (lattices[:, 1] * np.sin(alphas))
+    orth[:, 2, 1] = -1 * np.cos(alphas) / (lattices[:, 2] * np.sin(alphas))
+    orth[:, 2, 2] = 1 / lattices[:, 2]
+
+    return orth
+
+
+def cell_vols(lattices: np.ndarray) -> np.ndarray:
+    """Calculates the unit-cell volumes of a series of N unit cells.
+
+    Parameters:
+        lattices: An N x 6 array of lattice parameters (a, b, c, α, β, γ)
+            in Å and degrees, respectively.
+
+    Returns:
+        The unit cell volumes in Å³.
+    """
+    alphas, betas, gammas = (
+        np.radians(lattices[:, 3]),
+        np.radians(lattices[:, 4]),
+        np.radians(lattices[:, 5]),
+    )
+    return np.product(lattices[:, :3], axis=1) * np.sqrt(
+        (1 - np.cos(alphas) ** 2 - np.cos(betas) ** 2 - np.cos(gammas) ** 2)
+        + (2 * np.cos(alphas) * np.cos(betas) * np.cos(gammas))
+    )
+
+
+def empirical_p_T(
+    p_T: np.ndarray, ε_0: np.ndarray, λ_P: float, p_c: float, ν: float
+) -> np.ndarray:
+    """Empirical fit for pressure input data.
+
+    Parameters:
+        p_T: p(T) array of pressure data points,
+        ε_0: strain at critical pressure
+        λ_P: compressibility in (GPa^-nu)
[…]
