+++ tests/test_app.py
@@ -4,7 +4,6 @@ on common inputs to the app.
 """
 
 from functools import partial
-from pathlib import Path
 
 from PASCal.app import app
 
@@ -20,30 +19,6 @@ def client():
 @fixture(scope="session")
 def parser():
     return partial(BeautifulSoup, features="html.parser")
-
-
-def load_sample_data(filename):
-    with open(
-        Path(__file__).parent / "data" / filename, "r", encoding="utf-8"
-    ) as f:
-        flines = f.readlines()
-    return "\r".join(flines).replace("\n\r", "\r\n").strip("\n")
-
-
-@fixture
-def sample_T():
-    return load_sample_data("sample_T.tsv")
-
-
-@fixture
-def sample_P():
-    return load_sample_data("sample_P.tsv")
-
-
-@fixture
-def sample_q():
-    return load_sample_data("sample_q.tsv")
-
 
 def test_T_sample_data(client, sample_T, parser):
     post_parameters = {
