### 2023-08-23 1350f62
Add e2e tests for all tables of output results in sample data
