+++ tests/test_app.py
@@ -16,11 +16,33 @@ def client():
     with app.test_client() as cli:
         yield cli
 
+
 @fixture(scope="session")
 def parser():
     return partial(BeautifulSoup, features="html.parser")
 
-def test_T_sample_data(client, sample_T, parser):
+
+def check_tables(table, expected_table):
+    """Check tables but allow rounding from 0.0 to -0.0."""
+    for ind, _ in enumerate(table):
+        if expected_table[ind] == "-0.0":
+            expected_table[ind] = "0.0"
+        if table[ind] == "-0.0":
+            table[ind] = "0.0"
+
+    assert table == expected_table
+    return True
+
+
+def test_T_sample_data(
+    client,
+    sample_T,
+    parser,
+    sample_T_output,
+    sample_T_delta_length,
+    sample_T_volumes,
+    sample_T_principal_axes,
+):
     post_parameters = {
         "DataType": "Temperature",
         "PcVal": "",
@@ -41,17 +63,43 @@ def test_T_sample_data(client, sample_T, parser):
     tables = soup.find_all("table")
     assert len(tables) == 5
 
+    assert check_tables(
+        [td.text for td in tables[0].find_all("td")], sample_T_output
+    ), "Output table failed"
 
-def test_P_sample_data(client, sample_P, parser):
+    assert check_tables(
+        [td.text for td in tables[1].find_all("td")], sample_T_delta_length
+    ), "% length table failed"
+
+    assert check_tables(
+        [td.text for td in tables[2].find_all("td")], sample_T_principal_axes
+    ), "Principal axes table failed"
+
+    assert check_tables(
+        [td.text for td in tables[3].find_all("td")], sample_T_volumes
+    ), "Volume table failed"
+
+
+def test_p_sample_data(
+    client,
+    sample_p,
+    parser,
+    sample_p_output,
+    sample_p_bm,
+    sample_p_K,
+    sample_p_delta_length,
+    sample_p_principal_axes,
+    sample_p_volumes,
+):
     post_parameters = {
         "DataType": "Pressure",
         "EulerianStrain": "True",
         "FiniteStrain": "True",
         "DegPolyCap": "",
         "DegPolyVol": "",
-        "Pc": "True",
-        "PcVal": 0.19,
-        "data": sample_P,
+        "UsePc": "True",
+        "PcVal": "0.19",
+        "data": sample_p,
     }
 
     response = client.post("/output", data=post_parameters)
@@ -64,13 +112,47 @@ def test_P_sample_data(client, sample_P, parser):
     tables = soup.find_all("table")
     assert len(tables) == 7
 
+    assert check_tables(
+        [td.text for td in tables[0].find_all("td")], sample_p_output
+    ), "Output table failed"
 
-def test_q_sample_data(client, sample_q, parser):
+    assert check_tables(
+        [td.text for td in tables[1].find_all("td")], sample_p_bm
+    ), "BM table failed"
+
+    assert check_tables(
+        [td.text for td in tables[2].find_all("td")], sample_p_K
+    ), "Compressibility table failed"
+
+    assert check_tables(
+        [td.text for td in tables[3].find_all("td")], sample_p_delta_length
+    ), "% length table failed"
+
+    assert check_tables(
+        [td.text for td in tables[4].find_all("td")], sample_p_principal_axes
+    ), "Principal axes table failed"
+
+    assert check_tables(
+        [td.text for td in tables[5].find_all("td")], sample_p_volumes
+    ), "Volume table failed"
+
+
+def test_q_sample_data(
+    client,
+    sample_q,
+    parser,
+    sample_q_output,
+    sample_q_volumes,
+    sample_q_principal_axes,
+    sample_q_charge_derivative,
[…]
