### 2023-08-23 e7787f2
Significantly refactor app: first pass up to templating
