+++ src/PASCal/constants.py
@@ -1,0 +1,4 @@
+PERCENT = 1e2
+GPa_to_TPa = 1e3  # GPa to TPa
+K_to_MK = 1e6  # K to MK
+mAhg_to_kAhg = 1e6  # mAhg-1 to kAhg-1
