### 2023-08-23 dd7efac
Significantly refactor app: first pass up to templating
