+++ src/PASCal/app.py
@@ -1300,12 +1300,12 @@ def fit(x, x_errors, unit_cells, options, raw_data):
             InputHeadings=InputHeadings,
             data=raw_data,
             Axes=Axes,
-            PrinComp=PASCal._legacy.Round(PrinComp, 4),
+            PrinComp=np.round(PrinComp, 4),
             CalAlphaErr=PASCal._legacy.Round(CalAlphaErr * K_to_MK, 4),
             MedianPrinAxCryst=PASCal._legacy.Round(MedianPrinAxCryst, 4),
             PrinAxCryst=PASCal._legacy.Round(PrinAxCryst, 4),
             TPx=x,
-            DiagStrain=PASCal._legacy.Round(diagonal_strain * PERCENT, 4),
+            DiagStrain=np.round(diagonal_strain * PERCENT, 4),
             XCal=PASCal._legacy.Round(XCal, 4),
             Vol=PASCal._legacy.Round(cell_volumes, 4),
             VolLin=PASCal._legacy.Round(VolLin, 4),
