### 2023-08-23 dedc2a1
Fix bug in option parsing
