### 2023-08-23 efbb641
Trigger CI runs on PRs but forgo test deployment
