+++ docs/api_reference.md
@@ -1,0 +1,2 @@
+title: API Reference
+::: PASCal
