+++ pyproject.toml
@@ -31,7 +31,12 @@ scipy = "*"
 statsmodels = "*"
 pytest = { version = "*", optional = true }
 beautifulsoup4 = { version = "*", optional = true }
+mkdocs = { version = "*", optional = true }
+mkdocs-material = { version = "*", optional = true }
+mkdocstrings = { version = "*", optional = true, extras = ["python-legacy"] }
+mkdocs-awesome-pages-plugin = { version = "*", optional = true }
 
 
 [tool.poetry.extras]
 test = ["pytest", "beautifulsoup4"]
+docs = ["mkdocs", "mkdocs-material", "mkdocstrings", "mkdocs-awesome-pages-plugin"]
