### 2023-08-24 ae2a4a5
Remove cruft from workflow
