+++ .github/workflows/ci.yml
@@ -59,13 +59,17 @@ jobs:
           !site/
           !docs/
           !tests/
+
+    - name: Fix permissions # followng https://github.com/actions/upload-pages-artifact?tab=readme-ov-file#file-permissions
+      run: |
+        chmod -c -R +rX "site/" | while read line; do
+          echo "::warning title=Invalid file permissions automatically fixed::$line"
+        done
     
     - name: Upload docs artifact 
-      uses: actions/upload-artifact@v3
+      uses: actions/upload-pages-artifact@v2
       with:
-        name: github-pages
-        path: |
-          ./site/
+        path: site
 
   deploy:
     if: ${{github.repository == 'MJCliffe/PASCal' && github.event_name == 'push' }}
