### 2023-08-27 91f15e4
Additional refactoring
