### 2023-08-27 a5e2abf
Additional refactoring
