+++ src/PASCal/app.py
@@ -1,19 +1,15 @@
 from typing import Tuple, List
 import json
 import os
-import itertools
 
 import PASCal._legacy
 import PASCal.utils
+from PASCal.fitting import fit_linear_wls, fit_empirical, fit_chebyshev
 from PASCal.constants import PERCENT, K_to_MK, GPa_to_TPa, mAhg_to_kAhg
 from PASCal.options import Options, PASCalDataType
+from PASCal.plotting import plot_strain, plot_volume
 
 from flask import Flask, render_template, request, send_from_directory
-import plotly
-import plotly.graph_objs as go
-import plotly.subplots
-import statsmodels.api as sm
-
 
 import numpy as np
 from scipy.optimize import curve_fit
@@ -144,23 +140,9 @@ def fit(x, x_errors, unit_cells, options, raw_data):
     median_x = int(np.ceil(len(x) / 2)) - 1  # median
 
     ### Axes matching
-    for n in range(2, len(unit_cells)):
-        Match = np.dot(
-            principal_axes[1, :].T, principal_axes[n, :]
-        )  # an array matching the axes against each other
-        AxesOrder = np.zeros(
-            (3), dtype=np.int8
-        )  # a list of the axes needed to convert the eigenvalues and vectors into a consistent format
-        TestAxesOrder = np.zeros((6))  # cost function for each item
-        permute = list(itertools.permutations([0, 1, 2]))  # all orders
-        m = 0
-        for item in permute:
-            for i in range(3):
-                TestAxesOrder[m] += np.abs(Match[i][item[i]])
-            m = m + 1
-        AxesOrder = np.array(permute[np.argmax(TestAxesOrder)])
-        diagonal_strain[n, [0, 1, 2]] = diagonal_strain[n, AxesOrder]
-        principal_axes[n, :, [0, 1, 2]] = principal_axes[n, :, AxesOrder]
+    principal_axes, diagonal_strain = PASCal.utils.match_axes(
+        principal_axes, unit_cells, diagonal_strain
+    )
 
     ### Calculating Eigenvectors and Cells in different coordinate systems
     PrinAxCryst = np.transpose(
@@ -172,222 +154,29 @@ def fit(x, x_errors, unit_cells, options, raw_data):
     PrinAxCryst = (
         PrinAxCryst.T / (np.sum(PrinAxCryst**2, axis=2) ** 0.5).T
     ).T  # normalised to make UVW near 1
+
     ### Ensures the largest component of each eigenvector is positive to make comparing easier
     MaxAxis = np.argmax(
         np.abs(PrinAxCryst), axis=2
     )  # find the largest value of each eigenvector
     I, J = np.indices(MaxAxis.shape)
     Mask = PrinAxCryst[I, J, MaxAxis] < 0
-    PrinAxCryst[Mask, :] = PrinAxCryst[Mask, :] * -1
+    PrinAxCryst[Mask, :] *= -1
     # transpositions to take advantage of broadcasting, not maths
     MedianPrinAxCryst = PrinAxCryst[median_x]
 
-    ####
     ## Linear fitting of volume and lattice parameters
-    ## Linear strain fitting
-    CalAlpha = np.zeros(3)
-    CalAlphaErr = np.zeros(3)
-    CalYInt = np.zeros(3)
-    XCal = np.zeros((3, len(x)))
-    for i in range(3):
-        X = sm.add_constant(x)
 
-        StrainModel = sm.WLS(
-            diagonal_strain[:, i], X, weights=1 / x_errors
-        )  # weighted least square
-        StrainResults = StrainModel.fit()
-        CalAlpha[i] = StrainResults.params[1]
[…]

[…]
