### 2023-08-28 ab20609
Significant refactors of plotting and fitting
