+++ src/PASCal/plotting.py
@@ -1,4 +1,5 @@
-from typing import Dict, List
+from re import L
+from typing import Dict, List, Union
 import json
 
 import plotly
@@ -20,7 +21,11 @@ PLOT_STRAIN_FIT_LABEL: List[str] = [
     "ε<sub>2,calc</sub>",
     "ε<sub>3,calc</sub>",
 ]
-
+PLOT_INDICATRIX_LABELS: Dict[PASCalDataType, str] = {
+    PASCalDataType.TEMPERATURE: "Expansivity (MK<sup>-1</sup>)",
+    PASCalDataType.PRESSURE: "K (TPa<sup>-1</sup>)",
+    PASCalDataType.ELECTROCHEMICAL: "Electrochemical strain charge derivative ([kAhg<sup>-1</sup>]<sup>-1</sup>)",
+}
 PLOT_X_LABELS = {
     PASCalDataType.TEMPERATURE: "Temperature (K)",
     PASCalDataType.PRESSURE: "Pressure (GPa)",
@@ -32,8 +37,9 @@ def plot_strain(
     x: np.ndarray,
     diagonal_strain: np.ndarray,
     fit_results: List[sm.regression.linear_model.RegressionResultsWrapper],
-    data_type: PASCalDataType,
-) -> str:
+    data_type: Union[PASCalDataType, str],
+    return_json: bool = False,
+) -> Union[go.Figure, str]:
     """Plot strain (%) against the fitted data in the 3 principal directions.
 
     Parameters:
@@ -43,7 +49,7 @@ def plot_strain(
         data_type: The PASCalDataType corresponding to the identity of the control variable.
 
     Returns:
-        JSON dump of the plotly figure.
+        JSON dump of the plotly figure if return_json, else return the plotly figure.
 
     """
 
@@ -52,8 +58,11 @@ def plot_strain(
             "Expected 3 fitted models for strain plot, received {fit_results=}"
         )
 
-    gradient = [_.params[1] for _ in fit_results]
+    if not isinstance(data_type, PASCalDataType):
+        data_type = PASCalDataType[data_type.upper()]
+
     intercept = [_.params[0] for _ in fit_results]
+    gradient = [_.params[1] for _ in fit_results]
 
     figure = go.Figure()
     for i in range(3):
@@ -71,7 +80,7 @@ def plot_strain(
         figure.add_trace(
             go.Scatter(
                 x=x,
-                y=(gradient[i] * x + cal_intercept[i]) * PERCENT,
+                y=(gradient[i] * x + intercept[i]) * PERCENT,
                 name=PLOT_STRAIN_FIT_LABEL[i],
                 mode="lines",
                 line=dict(color=PLOT_PALETTE[i]),
@@ -104,15 +113,18 @@ def plot_strain(
         plot_bgcolor="white",
     )
 
-    return json.dumps(figure, cls=plotly.utils.PlotlyJSONEncoder)
+    if return_json:
+        json.dumps(figure, cls=plotly.utils.PlotlyJSONEncoder)
+    return figure
 
 
 def plot_volume(
     x: np.ndarray,
     cell_volumes: np.ndarray,
     fit_result: sm.regression.linear_model.RegressionResultsWrapper,
-    data_type: PASCalDataType,
-) -> str:
+    data_type: Union[PASCalDataType, str],
+    return_json: bool = False,
+) -> Union[go.Figure, str]:
     """Plot the cell volume fit.
 
     Parameters:
@@ -120,11 +132,15 @@ def plot_volume(
         cell_volumes: array of cell volumes.
         fit_results: A statsmodel linear regression result for the cell volume fit.
         data_type: The PASCalDataType corresponding to the identity of the control variable.
[…]
