+++ src/PASCal/options.py
@@ -1,6 +1,7 @@
 from dataclasses import dataclass, field, fields
-from typing import Optional, Dict
+from typing import Optional, Dict, Any
 from enum import Enum
+
 
 class PASCalDataType(Enum):
     PRESSURE = "Pressure"
@@ -10,6 +11,12 @@ class PASCalDataType(Enum):
 
 @dataclass
 class Options:
+    data_type: PASCalDataType = field(
+        metadata={
+            "form": "DataType",
+            "description": "The type of data passed, either 'Pressure', 'Electrochemical' or 'Temperature'",
+        },
+    )
     eulerian_strain: bool = field(
         default=True,
         metadata={
@@ -18,13 +25,6 @@ class Options:
         },
     )
     finite_strain: bool = field(default=True, metadata={"form": "FiniteStrain"})
-    data_type: PASCalDataType = field(
-        default=None,
-        metadata={
-            "form": "DataType",
-            "description": "The type of data passed, either 'Pressure', 'Electrochemical' or 'Temperature'",
-        },
-    )
     use_pc: bool = field(
         default=False,
         metadata={
@@ -39,41 +39,47 @@ class Options:
             "description": "Critical pressure value to use in GPa",
         },
     )
-    deg_poly_cap: Optional[int] = field(
+    deg_poly_strain: Optional[int] = field(
         default=None,
         metadata={
             "form": "DegPolyCap",
-            "description": "The degree of polynomial to use for fitting the capacity.",
+            "description": "The degree of polynomial to use for fitting the strain vs charge/capacity.",
         },
     )
     deg_poly_vol: Optional[int] = field(
         default=None,
         metadata={
             "form": "DegPolyVol",
-            "description": "The degree of polynomial to use for fitting the volume.",
+            "description": "The degree of polynomial to use for fitting the volume vs charge/capacity",
         },
     )
 
     @staticmethod
-    def from_form(self, form_data: Dict[str, str]) -> Options:
-    """Go through the set values of options and check their values."""
+    def from_dict(options: Dict[str, Any]) -> "Options":
+        """Load options from a dictionary."""
 
-    options = {}
+        if options.get("data_type"):
+            options["data_type"] = PASCalDataType[options["data_type"].upper()]
+        if options.get("pc_val"):
+            options["pc_val"] = float(options["pc_val"])
+        if options.get("deg_poly_strain"):
+            options["deg_poly_strain"] = int(options["deg_poly_strain"])
+        if options.get("deg_poly_vol"):
+            options["deg_poly_vol"] = int(options["deg_poly_vol"])
 
-    for key in fields(Options):
-        form_key = key.metadata["form"]
-        if form_key in form_data:
-            value = True if form_data[form_key] == "True" else form_data[form_key]
-            value = False if form_data[form_key] == "False" else value
-            options[key.name] = value
+        return Options(**options)
 
-    if options.get("data_type"):
-        options["data_type"] = PASCalDataType(options["data_type"])
-    if options.get("pc_val"):
-        options["pc_val"] = float(options["pc_val"])
-    if options.get("deg_poly_cap"):
-        options["deg_poly_cap"] = int(options["deg_poly_cap"])
-    if options.get("deg_poly_vol"):
-        options["deg_poly_vol"] = int(options["deg_poly_vol"])
+    @staticmethod
+    def from_form(form_data: Dict[str, str]) -> "Options":
+        """Go through and check the values provided by the form."""
 
-    return Options(**options)
+        options = {}
+
+        for key in fields(Options):
+            form_key = key.metadata["form"]
+            if form_key in form_data:
+                value = True if form_data[form_key] == "True" else form_data[form_key]
+                value = False if form_data[form_key] == "False" else value
+                options[key.name] = value
+
+        return Options.from_dict(options)

[…]
