### 2023-08-28 13a22e7
Significant refactors of plotting and fitting
