+++ src/PASCal/app.py
@@ -1,18 +1,21 @@
-from typing import Tuple, List
+from typing import Tuple, List, Optional, Union
 import json
 import os
 
 import PASCal._legacy
 import PASCal.utils
-from PASCal.fitting import fit_linear_wls, fit_empirical, fit_chebyshev
+from PASCal.fitting import (
+    fit_linear_wls,
+    fit_empirical_strain_pressure,
+    fit_birch_murnaghan_volume_pressure,
+    fit_chebyshev,
+)
 from PASCal.constants import PERCENT, K_to_MK, GPa_to_TPa, mAhg_to_kAhg
 from PASCal.options import Options, PASCalDataType
-from PASCal.plotting import plot_strain, plot_volume
 
 from flask import Flask, render_template, request, send_from_directory
 
 import numpy as np
-from scipy.optimize import curve_fit
 
 app = Flask(__name__)
 
@@ -70,7 +73,9 @@ def output():
             f"Could not parse data: {request.form.get('data')}\nException: {exc}"
         )
 
-    return fit(x, x_errors, unit_cells, options, raw_data)
+    fit_results = fit(x, x_errors, unit_cells, options, raw_data)
+
+    return _render_page(fit_results, options)
 
 
 def _precheck_inputs(x, options) -> Tuple[Options, List[str]]:
@@ -102,14 +107,14 @@ def _precheck_inputs(x, options) -> Tuple[Options, List[str]]:
                 options.pc_val = pc_val
 
     if options.data_type == PASCalDataType.ELECTROCHEMICAL:
-        if len(x) - 2 < options.deg_poly_cap:
-            deg_poly_cap = len(x) - 2
+        if len(x) - 2 < options.deg_poly_strain:
+            deg_poly_strain = len(x) - 2
             warning.append(
-                f"The maximum degree of the Chebyshev strain polynomial has been lowered from {options.deg_poly_cap} to {deg_poly_cap}. "
+                f"The maximum degree of the Chebyshev strain polynomial has been lowered from {options.deg_poly_strain} to {deg_poly_strain}. "
                 "At least as many data points as parameters are needed for a fit to be carried out. "
                 "As PASCal calculates errors from derivatives, more data points than parameters are needed for error estimates."
             )
-            options.deg_poly_cap = deg_poly_cap
+            options.deg_poly_strain = deg_poly_strain
         if len(x) - 2 < options.deg_poly_vol:
             deg_poly_vol = len(x) - 2
             warning.append(
@@ -122,9 +127,12 @@ def _precheck_inputs(x, options) -> Tuple[Options, List[str]]:
     return options, warning
 
 
-def fit(x, x_errors, unit_cells, options, raw_data):
+def fit(x, x_errors, unit_cells, options: Union[Options, dict]):
+    if not isinstance(options, Options):
+        options = Options.from_dict(options)
+
     options, warning = _precheck_inputs(x, options)
-    cell_volumes = PASCal._legacy.CellVol(unit_cells)  # volumes in (Angstrom^3)
+    cell_volumes = PASCal._legacy.CellVol(unit_cells)
 
     orthonormed_cells = PASCal._legacy.Orthomat(unit_cells)  # cell in orthogonal axes
 
@@ -145,866 +153,69 @@ def fit(x, x_errors, unit_cells, options, raw_data):
     )
 
     ### Calculating Eigenvectors and Cells in different coordinate systems
-    PrinAxCryst = np.transpose(
-        np.matmul(orthonormed_cells, principal_axes[:, :, :]), axes=[0, 2, 1]
-    )  # Eigenvector projected on crystallographic axes, UVW
-    CrystPrinAx = np.linalg.inv(
-        PrinAxCryst
-    )  # Unit Cell in Principal axes coordinates, å
[…]
