### 2023-08-28 8430327
Create core PASCal module
