### 2023-08-28 77d77b7
Create core PASCal module
