+++ src/PASCal/core.py
@@ -1,0 +1,174 @@
+from dataclasses import dataclass
+from typing import Union, List, Tuple
+from PASCal.options import Options, PASCalDataType
+from PASCAl.utils import Strain, Pressure, Volume, Charge, Temperature
+from PASCal.constants import PERCENT, K_to_MK, GPa_to_TPa, mAhg_to_kAhg
+from PASCal.fitting import (
+    fit_linear_wls,
+    fit_chebyshev,
+    fit_birch_murnaghan_volume_pressure,
+    fit_empirical_strain_pressure,
+)
+import statsmodel.api as sm
+import PASCal.utils
+import numpy as np
+
+
+class PASCalResults(dataclass):
+    """A container for the results of a PASCal run."""
+
+    options: Options
+    x: Union[Pressure, Temperature, Charge]
+    x_error: Union[Pressure, Temperature, Charge]
+    diagonal_strain: Strain
+    cell_volumes: Volume
+    prinicipal_components: np.ndarray
+    indicatrix: Tuple
+    norm_crax: np.ndarray
+    median_x: int
+    strain_fits: List[sm.regression.linear_model.RegressionResultsWrapper, np.ndarray]
+    volume_fits: List[sm.regression.linear_model.RegressionResultsWrapper, np.ndarray]
+
+
+def _precheck_inputs(x, options) -> Tuple[Options, List[str]]:
+    """Check that the raw data passed is compatible with the options, adjusting
+    the options where possible.
+
+    Returns:
+        The adjusted options and a list of warnings.
+
+    """
+    if len(x) < 2:
+        raise RuntimeError("Too few data points to perform fit: need at least 2")
+
+    warning: List[str] = []
+
+    if options.data_type == PASCalDataType.PRESSURE:
+        if len(x) < 4:
+            warning.append(
+                "At least as many data points as parameters are needed for a fit to be carried out (e.g. 3 for 3rd order Birch-Murnaghan, 4 for empirical pressure fitting). "
+                "As PASCal calculates errors from derivatives, more data points than parameters are needed for error estimates."
+            )
+        if options.use_pc and options.pc_val:
+            if np.amin(x) < options.pc_val:
+                pc_val = np.min(x)
+                warning.append(
+                    "The critical pressure has to be smaller than the lower pressure data point. "
+                    f"Critical pressure has been set to the minimum value: {pc_val} GPa."
+                )
+                options.pc_val = pc_val
+
+    if options.data_type == PASCalDataType.ELECTROCHEMICAL:
+        if len(x) - 2 < options.deg_poly_strain:
+            deg_poly_strain = len(x) - 2
+            warning.append(
+                f"The maximum degree of the Chebyshev strain polynomial has been lowered from {options.deg_poly_strain} to {deg_poly_strain}. "
+                "At least as many data points as parameters are needed for a fit to be carried out. "
+                "As PASCal calculates errors from derivatives, more data points than parameters are needed for error estimates."
+            )
+            options.deg_poly_strain = deg_poly_strain
+        if len(x) - 2 < options.deg_poly_vol:
+            deg_poly_vol = len(x) - 2
+            warning.append(
+                f"The maximum degree of the Chebyshev volume polynomial has been lowered from {options.deg_poly_vol} to {deg_poly_vol}. "
+                "At least as many data points as parameters are needed for a fit to be carried out. "
+                "As PASCal calculates errors from derivatives, more data points than parameters are needed for error estimates."
+            )
+            options.deg_poly_vol = deg_poly_vol
+
[…]
