+++ src/PASCal/app.py
@@ -1,16 +1,8 @@
-from typing import Tuple, List, Optional, Union
-import json
+from typing import Tuple
 import os
 
 import PASCal._legacy
-import PASCal.utils
-from PASCal.fitting import (
-    fit_linear_wls,
-    fit_empirical_strain_pressure,
-    fit_birch_murnaghan_volume_pressure,
-    fit_chebyshev,
-)
-from PASCal.constants import PERCENT, K_to_MK, GPa_to_TPa, mAhg_to_kAhg
+from PASCal.core import fit
 from PASCal.options import Options, PASCalDataType
 
 from flask import Flask, render_template, request, send_from_directory
@@ -76,143 +68,6 @@ def output():
     fit_results = fit(x, x_errors, unit_cells, options, raw_data)
 
     return _render_page(fit_results, options)
-
-
-def _precheck_inputs(x, options) -> Tuple[Options, List[str]]:
-    """Check that the raw data passed is compatible with the options, adjusting
-    the options where possible.
-
-    Returns:
-        The adjusted options and a list of warnings.
-
-    """
-    if len(x) < 2:
-        raise RuntimeError("Too few data points to perform fit: need at least 2")
-
-    warning: List[str] = []
-
-    if options.data_type == PASCalDataType.PRESSURE:
-        if len(x) < 4:
-            warning.append(
-                "At least as many data points as parameters are needed for a fit to be carried out (e.g. 3 for 3rd order Birch-Murnaghan, 4 for empirical pressure fitting). "
-                "As PASCal calculates errors from derivatives, more data points than parameters are needed for error estimates."
-            )
-        if options.use_pc and options.pc_val:
-            if np.amin(x) < options.pc_val:
-                pc_val = np.min(x)
-                warning.append(
-                    "The critical pressure has to be smaller than the lower pressure data point. "
-                    f"Critical pressure has been set to the minimum value: {pc_val} GPa."
-                )
-                options.pc_val = pc_val
-
-    if options.data_type == PASCalDataType.ELECTROCHEMICAL:
-        if len(x) - 2 < options.deg_poly_strain:
-            deg_poly_strain = len(x) - 2
-            warning.append(
-                f"The maximum degree of the Chebyshev strain polynomial has been lowered from {options.deg_poly_strain} to {deg_poly_strain}. "
-                "At least as many data points as parameters are needed for a fit to be carried out. "
-                "As PASCal calculates errors from derivatives, more data points than parameters are needed for error estimates."
-            )
-            options.deg_poly_strain = deg_poly_strain
-        if len(x) - 2 < options.deg_poly_vol:
-            deg_poly_vol = len(x) - 2
-            warning.append(
-                f"The maximum degree of the Chebyshev volume polynomial has been lowered from {options.deg_poly_vol} to {deg_poly_vol}. "
-                "At least as many data points as parameters are needed for a fit to be carried out. "
-                "As PASCal calculates errors from derivatives, more data points than parameters are needed for error estimates."
-            )
-            options.deg_poly_vol = deg_poly_vol
-
-    return options, warning
-
-
-def fit(x, x_errors, unit_cells, options: Union[Options, dict]):
-    if not isinstance(options, Options):
-        options = Options.from_dict(options)
-
-    options, warning = _precheck_inputs(x, options)
-    cell_volumes = PASCal._legacy.CellVol(unit_cells)
-
-    orthonormed_cells = PASCal._legacy.Orthomat(unit_cells)  # cell in orthogonal axes
-
-    strain = PASCal.utils.calculate_strain(
[…]
