### 2023-08-28 871acfd
Wrap results in new PASCalResults class for easier plotting and rendering
