+++ src/PASCal/core.py
@@ -1,33 +1,60 @@
 from dataclasses import dataclass
-from typing import Union, List, Tuple
+from typing import Union, List, Tuple, Any, Optional
 from PASCal.options import Options, PASCalDataType
-from PASCAl.utils import Strain, Pressure, Volume, Charge, Temperature
-from PASCal.constants import PERCENT, K_to_MK, GPa_to_TPa, mAhg_to_kAhg
+from PASCal.plotting import plot_strain, plot_volume
+from PASCal.utils import Strain, Pressure, Volume, Charge, Temperature
+from PASCal.constants import K_to_MK, GPa_to_TPa
 from PASCal.fitting import (
     fit_linear_wls,
     fit_chebyshev,
     fit_birch_murnaghan_volume_pressure,
     fit_empirical_strain_pressure,
 )
-import statsmodel.api as sm
 import PASCal.utils
+import PASCal._legacy
 import numpy as np
 
 
-class PASCalResults(dataclass):
+@dataclass
+class PASCalResults:
     """A container for the results of a PASCal run."""
 
     options: Options
     x: Union[Pressure, Temperature, Charge]
-    x_error: Union[Pressure, Temperature, Charge]
+    x_errors: Union[Pressure, Temperature, Charge]
     diagonal_strain: Strain
     cell_volumes: Volume
-    prinicipal_components: np.ndarray
+    principal_components: np.ndarray
     indicatrix: Tuple
     norm_crax: np.ndarray
     median_x: int
-    strain_fits: List[sm.regression.linear_model.RegressionResultsWrapper, np.ndarray]
-    volume_fits: List[sm.regression.linear_model.RegressionResultsWrapper, np.ndarray]
+    strain_fits: List[Any]
+    volume_fits: List[Any]
+    compressibility: Optional[np.ndarray]
+    compressibility_errors: Optional[np.ndarray]
+    warning: List[str]
+
+    def plot_strain(self, return_json: bool = False, show_errors: bool = False):
+        return plot_strain(
+            self.x,
+            self.x_errors,
+            self.diagonal_strain,
+            self.strain_fits,
+            self.options.data_type,
+            return_json=return_json,
+            show_errors=show_errors,
+        )
+
+    def plot_volume(self, return_json: bool = False, show_errors: bool = False):
+        return plot_volume(
+            self.x,
+            self.x_errors,
+            self.cell_volumes,
+            self.volume_fits,
+            self.options.data_type,
+            return_json=return_json,
+            show_errors=show_errors,
+        )
 
 
 def _precheck_inputs(x, options) -> Tuple[Options, List[str]]:
@@ -84,6 +111,7 @@ def fit(x, x_errors, unit_cells, options: Union[Options, dict]):
         options = Options.from_dict(options)
 
     options, warning = _precheck_inputs(x, options)
+    print(f"Performing fit with {options=}")
     cell_volumes = PASCal._legacy.CellVol(unit_cells)
 
     orthonormed_cells = PASCal._legacy.Orthomat(unit_cells)  # cell in orthogonal axes
@@ -113,34 +141,34 @@ def fit(x, x_errors, unit_cells, options: Union[Options, dict]):
         orthonormed_cells, principal_axes, median_x
     )
 
-    ## Linear fitting of volume and lattice parameters
-    strain_linear_fits = fit_linear_wls(diagonal_strain, x, x_errors)
-    volume_linear_fits = fit_linear_wls(cell_volumes, x, x_errors)[0]
+    strain_fits = {}
+    volume_fits = {}
+
+    strain_fits["linear"] = fit_linear_wls(diagonal_strain, x, x_errors)
+    volume_fits["linear"] = fit_linear_wls(cell_volumes, x, x_errors)[0]
 
[…]
