+++ docs/api_reference/options.md
@@ -1,0 +1,1 @@
+::: PASCal.options
