+++ docs/api_reference/fitting.md
@@ -1,0 +1,1 @@
+::: PASCal.fitting
