+++ docs/api_reference/core.md
@@ -1,0 +1,1 @@
+::: PASCal.core
