+++ docs/api_reference.md
@@ -1,2 +1,0 @@
-title: API Reference
-::: PASCal
