### 2023-08-28 a536fe4
Move more into plotting module and update docstrings
