+++ src/PASCal/app.py
@@ -1,13 +1,17 @@
 from typing import Tuple
 import os
+import json
 
 import PASCal._legacy
-from PASCal.core import fit
+from PASCal.constants import K_to_MK, PERCENT
+from PASCal.core import PASCalResults, fit
 from PASCal.options import Options, PASCalDataType
 
 from flask import Flask, render_template, request, send_from_directory
 
 import numpy as np
+
+from PASCal.plotting import PLOTLY_CONFIG
 
 app = Flask(__name__)
 
@@ -65,119 +69,137 @@ def output():
             f"Could not parse data: {request.form.get('data')}\nException: {exc}"
         )
 
-    fit_results = fit(x, x_errors, unit_cells, options, raw_data)
+    fit_results = fit(x, x_errors, unit_cells, options)
 
-    return _render_page(fit_results, options)
+    return _render_page(fit_results)
 
 
-def _render_page(fit_results, options: Options):
-    if options.data_type == PASCalDataType.TEMPERATURE:
-        response = render_template(
+def _render_page(results: PASCalResults):
+    if results.options.data_type == PASCalDataType.TEMPERATURE:
+        return render_template(
             "temperature.html",
-            config=plotly_config,
-            warning=warning,
-            PlotStrainJSON=StrainJSON,
-            PlotVolumeJSON=VolumeJSON,
-            PlotIndicJSON=IndicatrixJSON,
-            CoeffThermHeadings=CoeffThermHeadings,
-            StrainHeadings=StrainHeadings,
-            VolTempHeadings=VolTempHeadings,
-            InputHeadings=InputHeadings,
-            data=raw_data,
-            Axes=Axes,
-            PrinComp=np.round(PrinComp, 4),
-            CalAlphaErr=PASCal._legacy.Round(CalAlphaErr * K_to_MK, 4),
-            MedianPrinAxCryst=PASCal._legacy.Round(median_principal_axis_crys, 4),
-            PrinAxCryst=PASCal._legacy.Round(principal_axis_crys, 4),
-            TPx=x,
-            DiagStrain=np.round(diagonal_strain * PERCENT, 4),
-            XCal=PASCal._legacy.Round(XCal, 4),
-            Vol=PASCal._legacy.Round(cell_volumes, 4),
-            VolLin=PASCal._legacy.Round(VolLin, 4),
-            VolCoef=PASCal._legacy.Round(VolCoef, 4),
-            VolCoefErr=PASCal._legacy.Round(VolCoefErr, 4),
-            TPxError=x_errors,
-            Latt=unit_cells,
-        )
-        return response
-
-    if options.data_type == PASCalDataType.PRESSURE:
-        return render_template(
-            "pressure.html",
-            config=plotly_config,
-            warning=warning,
-            PlotStrainJSON=StrainJSON,
-            PlotDerivJSON=DerivJSON,
-            PlotVolumeJSON=VolumeJSON,
-            PlotIndicJSON=IndicatrixJSON,
-            KEmpHeadings=KEmpHeadings,
-            CalEpsilon0=PASCal._legacy.Round(CalEpsilon0, 4),
-            CalLambda=PASCal._legacy.Round(CalLambda, 4),
-            CalPc=PASCal._legacy.Round(CalPc, 4),
-            CalNu=PASCal._legacy.Round(CalNu, 4),
-            StrainHeadings=StrainHeadings,
-            InputHeadings=InputHeadings,
-            data=raw_data,
-            Axes=Axes,
[…]
