+++ src/PASCal/core.py
@@ -1,7 +1,10 @@
+"""This module defines the core functionality of PASCal: the fit function and the results container."""
+
+import plotly.graph_objs
 from dataclasses import dataclass
-from typing import Union, List, Tuple, Any, Optional
+from typing import Union, List, Tuple, Any, Optional, Dict
 from PASCal.options import Options, PASCalDataType
-from PASCal.plotting import plot_strain, plot_volume
+from PASCal.plotting import plot_indicatrix, plot_strain, plot_volume
 from PASCal.utils import Strain, Pressure, Volume, Charge, Temperature
 from PASCal.constants import K_to_MK, GPa_to_TPa
 from PASCal.fitting import (
@@ -17,24 +20,70 @@ import numpy as np
 
 @dataclass
 class PASCalResults:
-    """A container for the results of a PASCal run."""
+    """A container for the results of a PASCal run, providing
+    convenience wrappers to all data objects and plotting functions.
+
+    """
 
     options: Options
-    x: Union[Pressure, Temperature, Charge]
-    x_errors: Union[Pressure, Temperature, Charge]
-    diagonal_strain: Strain
-    cell_volumes: Volume
-    principal_components: np.ndarray
-    indicatrix: Tuple
-    norm_crax: np.ndarray
-    median_x: int
-    strain_fits: List[Any]
-    volume_fits: List[Any]
-    compressibility: Optional[np.ndarray]
-    compressibility_errors: Optional[np.ndarray]
-    warning: List[str]
+    """The provided fit options."""
 
-    def plot_strain(self, return_json: bool = False, show_errors: bool = False):
+    x: Union[Pressure, Temperature, Charge]
+    """The input control variable $X$, either pressure, temperature or charge."""
+    x_errors: Union[Pressure, Temperature, Charge]
+    """The input errors on the control variable $X$, either pressure, temperature or charge."""
+
+    unit_cells: np.ndarray
+    """The input unit cells."""
+
+    diagonal_strain: Strain
+    """The diagonalised strain."""
+
+    cell_volumes: Volume
+    """The computed cell volumes."""
+
+    principal_components: np.ndarray
+    """The eigenvalues of the strain."""
+
+    indicatrix: Tuple
+    """The indicatrix data used to construct the surface plot."""
+
+    norm_crax: np.ndarray
+    """The normalised crystal axes."""
+
+    median_x: int
+    """The index of the middle value of the control variable $X$."""
+
+    median_principal_axis_crys: np.ndarray
+    """The median principal axis of the strain."""
+
+    principal_axis_crys: np.ndarray
+    """The principal axes of the strain."""
+
+    strain_fits: Dict[str, List[Any]]
+    """A dictionary containing the strain fits. Under each key is a list of fit models for each principal axis,
+    containing either a statsmodel fit result (temperature data), a tuple of `popt`, `pcov`, `callable` results from SciPy's `curve_fit` (pressure data),
+    or an array of Chebyshev coefficients from NumPy's `chebfit` (electrochemical data).
+    """
+    volume_fits: Dict[str, Any]
+    """A dictionary containing the volume fits. Under each key is a model for the cell volume,
+    containing either a statsmodel fit result, a tuple of `popt`, `pcov`, `callable` results from SciPy's `curve_fit`,
+    or an array of Chebyshev coefficients from NumPy's `chebfit`.
+    """
+
+    compressibility: Optional[np.ndarray]
+    """The computed compressibility."""
+
+    compressibility_errors: Optional[np.ndarray]
+    """The computed compressibility errors."""
+
+    warning: List[str]
+    """Any warnings generated during the fit."""
+
+    def plot_strain(
+        self, return_json: bool = False, show_errors: bool = False
+    ) -> Union[str, plotly.graph_objs.Figure]:
[…]
