### 2023-08-28 2e030cb
Move more into plotting module and update docstrings
