+++ src/PASCal/options.py
@@ -1,6 +1,10 @@
+"""This module defines the options used by the PASCal library/app
+when fitting strain and volume data.
+"""
 from dataclasses import dataclass, field, fields
-from typing import Optional, Dict, Any
+from typing import Optional, Dict, Any, List
 from enum import Enum
+import numpy as np
 
 
 class PASCalDataType(Enum):
@@ -14,45 +18,52 @@ class Options:
     data_type: PASCalDataType = field(
         metadata={
             "form": "DataType",
-            "description": "The type of data passed, either 'Pressure', 'Electrochemical' or 'Temperature'",
         },
     )
+    """The type of data passed, either 'Pressure', 'Electrochemical' or 'Temperature'."""
+
     eulerian_strain: bool = field(
         default=True,
         metadata={
             "form": "EulerianStrain",
-            "description": "Whether to use Eulerian strain (True) or Lagrangian",
         },
     )
+    """Whether to use Eulerian strain (True) or Lagrangian"""
+
     finite_strain: bool = field(default=True, metadata={"form": "FiniteStrain"})
+    """Whether to use finite strain (True) or infinitesimal strain"""
+
     use_pc: bool = field(
         default=False,
         metadata={
             "form": "Pc",
-            "description": "Whether to use the critical pressure to modify the fits.",
         },
     )
+    """Whether to use the critical pressure to modify the fits."""
+
     pc_val: Optional[float] = field(
         default=None,
         metadata={
             "form": "PcVal",
-            "description": "Critical pressure value to use in GPa",
         },
     )
+    """The critical pressure value to use in GPa."""
+
     deg_poly_strain: int = field(
         default=5,
         metadata={
             "form": "DegPolyCap",
-            "description": "The degree of polynomial to use for fitting the strain vs charge/capacity.",
         },
     )
+    """The degree of polynomial to use for fitting the strain vs charge/capacity."""
+
     deg_poly_vol: int = field(
         default=5,
         metadata={
             "form": "DegPolyVol",
-            "description": "The degree of polynomial to use for fitting the volume vs charge/capacity",
         },
     )
+    """The degree of polynomial to use for fitting the volume vs charge/capacity"""
 
     @staticmethod
     def from_dict(options: Dict[str, Any]) -> "Options":
@@ -60,11 +71,11 @@ class Options:
 
         if options.get("data_type"):
             options["data_type"] = PASCalDataType[options["data_type"].upper()]
-        if options.get("pc_val") is not None:
+        if options.get("pc_val"):
             options["pc_val"] = float(options["pc_val"])
-        if options.get("deg_poly_strain") is not None:
+        if options.get("deg_poly_strain"):
             options["deg_poly_strain"] = int(options["deg_poly_strain"])
-        if options.get("deg_poly_vol") is not None:
+        if options.get("deg_poly_vol"):
             options["deg_poly_vol"] = int(options["deg_poly_vol"])
 
         return Options(**options)
@@ -83,3 +94,51 @@ class Options:
                 options[key.name] = value
 
         return Options.from_dict(options)
+
+    def precheck_inputs(self, x) -> List[str]:
+        """Check that the raw data passed is compatible with the options, adjusting
+        the options where possible.
+
+        Returns:
+            The adjusted options and a list of warnings.
+
+        """
+        if len(x) < 2:
+            raise RuntimeError("Too few data points to perform fit: need at least 2")
+
+        warning: List[str] = []
+
+        if self.data_type == PASCalDataType.PRESSURE:
+            if len(x) < 4:
+                warning.append(
+                    "At least as many data points as parameters are needed for a fit to be carried out (e.g. 3 for 3rd order Birch-Murnaghan, 4 for empirical pressure fitting). "
[…]
