### 2023-08-28 831a396
Finished temperature template
