### 2023-08-28 dafe137
Finished temperature template
