+++ src/PASCal/app.py
@@ -71,10 +71,20 @@ def output():
 
     fit_results = fit(x, x_errors, unit_cells, options)
 
-    return _render_page(fit_results)
+    return _render_results(fit_results)
 
 
-def _render_page(results: PASCalResults):
+def _render_results(results: PASCalResults) -> str:
+    """Take the results of a PASCal fit and render them as HTML.
+
+    Parameters:
+        results: The results of a PASCal fit.
+
+    Returns:
+        The rendered HTML to serve.
+
+    """
+
     if results.options.data_type == PASCalDataType.TEMPERATURE:
         return render_template(
             "temperature.html",
@@ -85,89 +95,72 @@ def _render_page(results: PASCalResults):
             PlotIndicJSON=results.plot_indicatrix(return_json=True),
             Axes=["X1", "X2", "X3", "V"],
             PrinComp=np.round(results.principal_components, 4),
-            CalAlphaErr=PASCal._legacy.Round(
-                np.array(
-                    [
-                        results.strain_fits["linear"][i].HC0_se[1] * K_to_MK
-                        for i in range(3)
-                    ]
-                ),
-                4,
-            ),
             MedianPrinAxCryst=PASCal._legacy.Round(
                 results.median_principal_axis_crys, 4
             ),
+            Vol=PASCal.utils.round_array(results.cell_volumes, 4),
             PrinAxCryst=PASCal._legacy.Round(results.principal_axis_crys, 4),
             TPx=results.x,
             DiagStrain=np.round(results.diagonal_strain * PERCENT, 4),
-            XCal=PASCal._legacy.Round(results.x, 4),
-            Vol=PASCal._legacy.Round(results.cell_volumes, 4),
-            VolLin=PASCal._legacy.Round(
-                results.volume_fits["linear"].params[1] * results.x
-                + results.volume_fits["linear"].params[0],
-                4,
-            ),
-            VolCoef=PASCal._legacy.Round(
-                results.volume_fits["linear"].params[1]
-                / results.cell_volumes[0]
-                * K_to_MK,
-                4,
-            ),
-            VolCoefErr=PASCal._legacy.Round(
-                results.volume_fits["linear"].HC0_se[1]
-                / results.cell_volumes[0]
-                * K_to_MK,
-                4,
-            ),
             TPxError=results.x_errors,
             Latt=results.unit_cells,
+            **{
+                k: PASCal.utils.round_array(results.named_coefficients[k], 4)
+                for k in results.named_coefficients
+            },
         )
 
-    # if options.data_type == PASCalDataType.PRESSURE:
-    #     return render_template(
-    #         "pressure.html",
-    #         config=plotly_config,
-    #         warning=warning,
-    #         PlotStrainJSON=StrainJSON,
-    #         PlotDerivJSON=DerivJSON,
-    #         PlotVolumeJSON=VolumeJSON,
-    #         PlotIndicJSON=IndicatrixJSON,
-    #         KEmpHeadings=KEmpHeadings,
-    #         CalEpsilon0=PASCal._legacy.Round(CalEpsilon0, 4),
-    #         CalLambda=PASCal._legacy.Round(CalLambda, 4),
-    #         CalPc=PASCal._legacy.Round(CalPc, 4),
[…]
