+++ src/PASCal/plotting.py
@@ -7,8 +7,9 @@ import plotly.subplots
 import numpy as np
 
 from PASCal.options import PASCalDataType
-from PASCal.constants import PERCENT
+from PASCal.constants import PERCENT, GPa_to_TPa, mAhg_to_kAhg
 from PASCal.utils import (
+    get_compressibility,
     Pressure,
     Temperature,
     Charge,
@@ -20,12 +21,6 @@ PLOT_WIDTH: int = 500
 PLOT_HEIGHT: int = 500
 PLOT_MARGINS: Dict[str, float] = dict(t=50, b=50, r=50, l=50)
 PLOT_PALETTE: List[str] = ["Red", "Green", "Blue"]
-PLOT_STRAIN_LABEL: List[str] = ["ε<sub>1</sub>", "ε<sub>2</sub>", "ε<sub>3</sub>"]
-PLOT_STRAIN_FIT_LABEL: List[str] = [
-    "ε<sub>1,calc</sub>",
-    "ε<sub>2,calc</sub>",
-    "ε<sub>3,calc</sub>",
-]
 PLOT_INDICATRIX_LABELS: Dict[PASCalDataType, str] = {
     PASCalDataType.TEMPERATURE: "Expansivity (MK<sup>-1</sup>)",
     PASCalDataType.PRESSURE: "K (TPa<sup>-1</sup>)",
@@ -100,7 +95,7 @@ def plot_strain(
                 x=x,
                 y=diagonal_strain[:, i] * PERCENT,
                 error_x=dict(type="data", array=x_errors, visible=show_errors),
-                name=PLOT_STRAIN_LABEL[i],
+                name=f"ε<sub>{i}</sub>",
                 mode="markers",
                 marker_symbol="circle-open",
                 marker=dict(color=PLOT_PALETTE[i]),
@@ -111,7 +106,7 @@ def plot_strain(
             go.Scatter(
                 x=xs,
                 y=ys[i] * PERCENT,
-                name=PLOT_STRAIN_FIT_LABEL[i],
+                name=f"ε<sub>{i},calc</sub>",
                 mode="lines",
                 line=dict(color=PLOT_PALETTE[i]),
             )
@@ -239,6 +234,293 @@ def plot_volume(
         showline=True,
         linecolor="black",
     )
+    figure.update_layout(
+        autosize=False,
+        width=PLOT_WIDTH,
+        height=PLOT_HEIGHT,
+        margin=PLOT_MARGINS,
+        showlegend=True,
+        hovermode="x unified",
+        plot_bgcolor="white",
+    )
+
+    if return_json:
+        return json.dumps(figure, cls=plotly.utils.PlotlyJSONEncoder)
+    return figure
+
+
+def plot_compressibility(
+    x: Pressure,
+    compressibility: np.ndarray,
+    compressibility_errors: np.ndarray,
+    fit_results: Dict[str, Any],
+    data_type: Union[PASCalDataType, str],
+    return_json: bool = False,
+):
+    """Plot each component of the compressibility alongside
+    the fitted values.
+
+    Parameters:
+        x: Array containing the control variable.
+        compressibility: Array of compressibility values.
+        compressibility_errors: Array of compressibility errors.
+        fit_results: A set of `curve_fit` results for the compressibility fits in
+            each direction, with key `empirical`.
+        return_json: Whether or not to return the JSON dump of the figure.
+
+    Returns:
+        JSON dump of the plotly figure if return_json, else return the plotly figure.
+
+    """
+    figure = go.Figure()
+
+    if not isinstance(data_type, PASCalDataType):
+        data_type = PASCalDataType[data_type.upper()]
+    if not data_type == PASCalDataType.PRESSURE:
[…]
