+++ src/PASCal/app.py
@@ -3,7 +3,7 @@ import os
 import json
 
 import PASCal._legacy
-from PASCal.constants import K_to_MK, PERCENT
+from PASCal.constants import PERCENT
 from PASCal.core import PASCalResults, fit
 from PASCal.options import Options, PASCalDataType
 
@@ -61,7 +61,6 @@ def output():
     except Exception as exc:
         raise RuntimeError(f"Could not parse options: {request.form}\nException: {exc}")
 
-    raw_data = request.form.get("data")
     try:
         x, x_errors, unit_cells = _parse_data()
     except Exception as exc:
@@ -110,89 +109,61 @@ def _render_results(results: PASCalResults) -> str:
             },
         )
 
-    if results.options.data_type == PASCalDataType.PRESSURE:
-        pass
-        # return render_template(
-        #     "pressure.html",
-        #     config=json.dumps(PLOTLY_CONFIG),
-        #     warning=results.warning,
-        #     PlotStrainJSON=results.plot_strain(return_json=True),
-        #     PlotVolumeJSON=results.plot_volume(return_json=True),
-        #     PlotIndicJSON=results.plot_indicatrix(return_json=True),
-        #     PlotDerivJSON=DerivJSON,
-        #     CalEpsilon0=PASCal._legacy.Round(CalEpsilon0, 4),
-        #     CalLambda=PASCal._legacy.Round(CalLambda, 4),
-        #     CalPc=PASCal._legacy.Round(CalPc, 4),
-        #     CalNu=PASCal._legacy.Round(CalNu, 4),
-        #     Axes=["X1", "X2", "X3", "V"],
-        #     PrinComp=np.round(results.principal_components, 4),
-        #     MedianPrinAxCryst=PASCal._legacy.Round(
-        #         results.median_principal_axis_crys, 4
-        #     ),
-        #     PrinAxCryst=PASCal._legacy.Round(results.principal_axis_crys, 4),
-        #     BMOrder=["2nd", "3rd"] + ["3rd with Pc"] if results.options.use_pc else [],
-        #     B0=PASCal._legacy.Round(B0, 4),
-        #     SigB0=PASCal._legacy.Round(SigB0, 4),
-        #     V0=PASCal._legacy.Round(V0, 4),
-        #     SigV0=PASCal._legacy.Round(SigV0, 4),
-        #     BPrime=PASCal._legacy.Round(BPrime, 4),
-        #     SigBPrime=SigBPrime,
-        #     PcCoef=PASCal._legacy.Round(PcCoef, 4),
-        #     CalPress=PASCal._legacy.Round(CalPress, 4),
+    elif results.options.data_type == PASCalDataType.PRESSURE:
+        return render_template(
+            "pressure.html",
+            config=json.dumps(PLOTLY_CONFIG),
+            warning=results.warning,
+            PlotStrainJSON=results.plot_strain(return_json=True),
+            PlotVolumeJSON=results.plot_volume(return_json=True),
+            PlotIndicJSON=results.plot_indicatrix(return_json=True),
+            PlotDerivJSON=results.plot_compressibility(return_json=True),
+            Axes=["X1", "X2", "X3", "V"],
+            PrinComp=np.round(results.principal_components, 4),
+            MedianPrinAxCryst=PASCal._legacy.Round(
+                results.median_principal_axis_crys, 4
+            ),
+            PrinAxCryst=PASCal._legacy.Round(results.principal_axis_crys, 4),
+            BMOrder=["2nd", "3rd"] + ["3rd with Pc"] if results.options.use_pc else [],
+            TPxError=results.x_errors,
+            Latt=results.unit_cells,
+            K=PASCal._legacy.Round(results.compressibility, 4),
[…]
