+++ src/PASCal/core.py
@@ -4,9 +4,23 @@ import plotly.graph_objs
 from dataclasses import dataclass, field
 from typing import Union, List, Tuple, Any, Optional, Dict
 from PASCal.options import Options, PASCalDataType
-from PASCal.plotting import plot_indicatrix, plot_strain, plot_volume
-from PASCal.utils import Strain, Pressure, Volume, Charge, Temperature
-from PASCal.constants import K_to_MK, GPa_to_TPa, PERCENT
+from PASCal.plotting import (
+    plot_charge_derivative,
+    plot_indicatrix,
+    plot_residual,
+    plot_strain,
+    plot_volume,
+    plot_compressibility,
+)
+from PASCal.utils import (
+    Strain,
+    Pressure,
+    Volume,
+    Charge,
+    Temperature,
+    empirical_pressure_strain_relation,
+)
+from PASCal.constants import K_to_MK, GPa_to_TPa, PERCENT, mAhg_to_kAhg
 from PASCal.fitting import (
     fit_linear_wls,
     fit_chebyshev,
@@ -30,6 +44,7 @@ class PASCalResults:
 
     x: Union[Pressure, Temperature, Charge]
     """The input control variable $X$, either pressure, temperature or charge."""
+
     x_errors: Union[Pressure, Temperature, Charge]
     """The input errors on the control variable $X$, either pressure, temperature or charge."""
 
@@ -120,6 +135,39 @@ class PASCalResults:
             return_json=return_json,
         )
 
+    def plot_compressibility(
+        self, return_json: bool = False
+    ) -> Union[str, plotly.graph_objs.Figure]:
+        return plot_compressibility(
+            self.x,
+            self.compressibility,
+            self.compressibility_errors,
+            self.strain_fits,
+            self.options.data_type,
+            return_json=return_json,
+        )
+
+    def plot_charge_derivative(
+        self, return_json: bool = False
+    ) -> Union[str, plotly.graph_objs.Figure]:
+        return plot_charge_derivative(
+            self.x,
+            self.diagonal_strain,
+            self.strain_fits,
+            self.options.data_type,
+            return_json=return_json,
+        )
+
+    def plot_residual(
+        self, return_json: bool = False
+    ) -> Union[str, plotly.graph_objs.Figure]:
+        return plot_residual(
+            self.strain_fits,
+            self.volume_fits,
+            self.options.data_type,
+            return_json=return_json,
+        )
+
     def _set_named_coeffs(self):
         """Compute a series of named coefficients for displaying in the app."""
         self.named_coefficients = {}
@@ -133,7 +181,7 @@ class PASCalResults:
             )
 
             self.named_coefficients["VolCoef"] = (
-                self.volume_fits["linear"].params[1] / self.cell_volumes[0] * K_to_MK
+                -self.volume_fits["linear"].params[1] / self.cell_volumes[0] * K_to_MK
             )
 
             self.named_coefficients["VolCoefErr"] = (
@@ -148,6 +196,96 @@ class PASCalResults:
                     )
                     for i in range(3)
                 ]
+            )
+        if self.options.data_type == PASCalDataType.PRESSURE:
+            self.named_coefficients["VolCoef"] = (
+                self.volume_fits["linear"].params[1] / self.cell_volumes[0] * GPa_to_TPa
+            )
+
+            self.named_coefficients["VolCoefErr"] = (
+                self.volume_fits["linear"].HC0_se[1] / self.cell_volumes[0] * GPa_to_TPa
+            )
+            self.named_coefficients["CalEpsilon0"] = np.array(
+                [self.strain_fits["empirical"][0][i][0] for i in range(3)]
+            )
+            self.named_coefficients["CalLambda"] = np.array(
[…]
