### 2023-08-29 8dc22bc
Fix bullet list in README
