### 2023-08-29 a1eb088
Reintroduce loop over cheb-space and fix residual + indicatrix plots
