### 2023-08-29 140540c
Reintroduce loop over cheb-space and fix residual + indicatrix plots
