### 2023-08-29 409bef4
Fix compressibility plots
