### 2023-08-29 1130d06
Add missing P table parameters
