### 2023-08-29 93b3dd5
Allow string n/a through rounding function
