### 2023-08-29 b07001a
Remove final legacy code
