### 2023-08-29 f9cccd6
Remove final legacy code
