### 2023-08-29 4e84ee9
Add stub notebook example
