### 2023-08-31 0724fb1
Set deployment workflow to run on new GitHub releases
