+++ .github/workflows/deploy.yml
@@ -1,7 +1,9 @@
 name: Deploy to main server
 
 on:
-  workflow_dispatch:
+  release:
+    types:
+     - published
 
 jobs:
 
