### 2023-08-31 f226ed5
Set the app version to 1.0.1 for final release before v2
