+++ src/PASCal/__init__.py
@@ -1,0 +1,5 @@
+import importlib.metadata
+
+__version__ = importlib.metadata.version("PASCal")
+
+__all__ = ("__version__",)
