### 2023-08-31 7d82668
Set the app version to 1.0.1 for final release before v2
