+++ pyproject.toml
@@ -1,10 +1,10 @@
 [build-system]
-requires = ["setuptools >= 42", "wheel"]
+requires = ["setuptools >= 61.0.0", "wheel"]
 build-backend = "setuptools.build_meta"
 
 [tool.poetry]
 name = "PASCal"
-version = "2.0.0"
+version = "1.0.1"
 description = "Principal Axis Strain Calculator (PASCal) is a web tool designed to help scientists analyse non-ambient lattice parameter data."
 readme = "README.md"
 keywords = ["materials", "chemistry", "fitting", "strain", "lattice"]
@@ -12,7 +12,7 @@ license = "MIT"
 authors = [ "Matthew Cliffe <matthew.cliffe@nottingham.ac.uk>", ]
 classifiers = [
     "Programming Language :: Python :: 3",
-    "Programming Language :: Python :: 3.10",
+    "Programming Language :: Python :: 3.8",
     "Intended Audience :: Science/Research",
     "Topic :: Scientific/Engineering",
 ]
