+++ paper.md
@@ -11,14 +11,19 @@
   authors:
     - name: Monthakan Lertkiattrakul
       affiliation: 1
+    - name: Matthew L. Evans
+      orcid:  0000-0002-1182-9098
+      affiliation: 2
     - name: Matthew J. Cliffe
       orcid: 0000-0002-0408-7647
-      affiliation: 1 
+      affiliation: 1
   affiliations:
-  - name: School of Chemistry, University Park, Nottingham, NG7 2RD, United Kingdom    
-    index: 1 
+  - name: School of Chemistry, University Park, Nottingham, NG7 2RD, United Kingdom
+    index: 1
+  - name: Institut de la Matière Condensée et des Nanosciences, Université catholique de Louvain, Chemin des Étoiles 8, Louvain-la-Neuve 1348, Belgium
+    index: 2
   date: 28 May 2023
-  
+
   bibliography: paper.bib
 ...
 
@@ -26,36 +31,32 @@
 
 # Summary
 
-The response of crystalline materials to external stimuli: whether temperature, pressure or electrochemical potential, is critical for both our understanding of materials and their use. This information can be readily obtained through in-situ diffraction experiments, however if the intrinsic anisotropy of crystals is not taken into account, the true behaviour of crystals can be overlooked. This is particularly true for anomalous mechanical properties of great topical interest such as negative linear or area compressibility,[@Cairns2015; @Hodgson2014] negative thermal expansion[@chenNegativeThermalExpansion2015] or strongly anisotropic electrochemical strain [@kondrakovAnisotropicLatticeStrain2017].
+The response of crystalline materials to external stimuli: whether temperature, pressure or electrochemical potential, is critical for both our understanding of materials and their use. This information can be readily obtained through in-situ diffraction experiments, however if the intrinsic anisotropy of crystals is not taken into account, the true behaviour of crystals can be overlooked. This is particularly true for anomalous mechanical properties of great topical interest, such as negative linear or area compressibility [@Cairns2015; @Hodgson2014], negative thermal expansion [@chenNegativeThermalExpansion2015] or strongly anisotropic electrochemical strain [@kondrakovAnisotropicLatticeStrain2017].
 
-Characterising this strain anisotropy requires calculating not just the deformation along unit cell axes directions, which for low symmetry crystals have an arbitrary orientation relative to the mechanical properties, but also how the changes in unit cell angles affect the principal (eigen-)strains and their orientations (eigenvectors). A great deal of additional information can be extracted from stimulus-dependence of these principal strains, and fitting this dependence to equations-of-state can produce parameters of fundamental importance (e.g. compressibilities). Software capable of transforming experimental lattice parameters into the principal strains and the derived underlying thermodynamic properties (e.g. expansivities or compressibilities) is therefore of great utility to experimental scientists studying the response of crystals.
+Characterising this strain anisotropy requires calculating not just the deformation along unit cell axes directions, which for low symmetry crystals have an arbitrary orientation relative to the mechanical properties, but also how changes in unit cell angles affect the principal (eigen-)strains and their orientations (eigenvectors). A great deal of additional information can be extracted from the stimulus-dependence of these principal strains, and fitting this dependence to equations-of-state can produce parameters of fundamental importance (e.g., compressibilities). Software capable of transforming experimental lattice parameters into the principal strains and the derived underlying thermodynamic properties (e.g. expansivities or compressibilities) is therefore of great utility to experimental scientists studying the response of crystals.
 
 # Statement of Need
 
[…]
