### 2023-08-31 e763883
Add version and style tweaks to templates
