### 2023-09-18 536c47d
Fix a couple of typos
