+++ pyproject.toml
@@ -4,7 +4,7 @@ build-backend = "setuptools.build_meta"
 
 [tool.poetry]
 name = "PASCal"
-version = "1.0.1"
+version = "2.0.0"
 description = "Principal Axis Strain Calculator (PASCal) is a web tool designed to help scientists analyse non-ambient lattice parameter data."
 readme = "README.md"
 keywords = ["materials", "chemistry", "fitting", "strain", "lattice"]
