### 2023-09-18 0dc4bb6
Pass plot size through indicatrix wrapper
