### 2023-09-18 7d0d010
Add Jupyter fitting examples to docs
