### 2023-09-20 db6819e
Fix visual bug in BM coeffs table
