### 2023-09-20 b76bc36
Add more informative warning and remove to-do
