### 2023-09-20 3256680
Add pre-commit to dev deps and CI and run various linters
