+++ .github/workflows/ci.yml
@@ -30,6 +30,10 @@ jobs:
         pip install poetry
         poetry install --all-extras
 
+    - name: Run linting
+      run: |
+        poetry run pre-commit run --all-files
+
     - name: Run tests
       run: |
         poetry run pytest
@@ -65,8 +69,8 @@ jobs:
         chmod -c -R +rX "site/" | while read line; do
           echo "::warning title=Invalid file permissions automatically fixed::$line"
         done
-    
-    - name: Upload docs artifact 
+
+    - name: Upload docs artifact
       uses: actions/upload-pages-artifact@v2
       with:
         path: site
@@ -106,7 +110,7 @@ jobs:
     environment:
       name: "Latest documentation"
       url: https://mjcliffe.github.io/PASCal
-    
+
     steps:
     - name: Deploy docs
       uses: actions/deploy-pages@v2
