### 2023-09-20 a335927
Add pre-commit to dev deps and CI and run various linters
