+++ mkdocs.yml
@@ -1,4 +1,4 @@
-site_name: PASCal 
+site_name: PASCal
 site_description: Documentation for PASCal
 
 theme:
