+++ .github/workflows/deploy.yml
@@ -9,7 +9,7 @@ jobs:
 
   build:
     runs-on: ubuntu-latest
-    steps:    
+    steps:
     - uses: actions/checkout@v3
       with:
         submodules: true
