### 2023-09-20 cd7abaf
Ignore spurious warnings in docs build
