### 2023-09-22 37dc9bf
included paper; swapped para in SoN & Summary
