+++ paper.bib
@@ -1,0 +1,227 @@
+@article{Birch1947,
+  title = {Finite {{Elastic Strain}} of {{Cubic Crystals}}},
+  author = {Birch, Francis},
+  year = {1947},
+  month = jun,
+  journal = {Physical Review},
+  volume = {71},
+  number = {11},
+  pages = {809--824},
+  issn = {0031-899X},
+  doi = {10.1103/PhysRev.71.809},
+  urldate = {2012-03-03}
+}
+% == BibTeX quality report for Birch1947:
+% ? Title looks like it was stored in title-case in Zotero
+% ? unused Url ("http://prola.aps.org/abstract/PR/v71/i11/p809_1")
+
+@article{Cairns2015,
+  title = {Negative Linear Compressibility},
+  author = {Cairns, Andrew B. and Goodwin, Andrew L.},
+  year = {2015},
+  month = feb,
+  journal = {Physical Chemistry Chemical Physics},
+  volume = {17},
+  number = {32},
+  eprint = {1502.00846},
+  pages = {20449--20465},
+  issn = {1463-9076},
+  doi = {10.1039/C5CP00442J},
+  abstract = {While all materials reduce their intrinsic volume under hydrostatic (uniform) compression, a select few actually expand along one or more directions during this process of densification.},
+  archiveprefix = {arxiv},
+  keywords = {Condensed Matter - Materials Science},
+  file = {/Users/pczmjc1/Library/CloudStorage/OneDrive-TheUniversityofNottingham/Papers/Zotero/2015/Physical Chemistry Chemical Physics/Cairns_Goodwin_2015_Negative linear compressibility.pdf}
+}
+% == BibTeX quality report for Cairns2015:
+% ? unused Url ("http://xlink.rsc.org/?DOI=C5CP00442J")
+
+@article{chenNegativeThermalExpansion2015,
+  title = {Negative Thermal Expansion in Functional Materials: Controllable Thermal Expansion by Chemical Modifications},
+  shorttitle = {Negative Thermal Expansion in Functional Materials},
+  author = {Chen, Jun and Hu, Lei and Deng, Jinxia and Xing, Xianran},
+  year = {2015},
+  journal = {Chemical Society Reviews},
+  volume = {44},
+  number = {11},
+  pages = {3522--3567},
+  publisher = {{Royal Society of Chemistry}},
+  doi = {10.1039/C4CS00461B},
+  urldate = {2023-04-02},
+  langid = {english},
+  file = {/Users/pczmjc1/Zotero/storage/XCDWYHZJ/Chen et al. - 2015 - Negative thermal expansion in functional materials.pdf}
+}
+% == BibTeX quality report for chenNegativeThermalExpansion2015:
+% ? unused Library catalog ("pubs.rsc.org")
+% ? unused Url ("https://pubs.rsc.org/en/content/articlelanding/2015/cs/c4cs00461b")
+
+@article{Cliffe2012b,
+  title = {{{PASCal}}: {{A}} Principal Axis Strain Calculator for Thermal Expansion and Compressibility Determination},
+  author = {Cliffe, Matthew J. and Goodwin, Andrew L.},
+  year = {2012},
+  month = nov,
+  journal = {Journal of Applied Crystallography},
+  volume = {45},
+  number = {6},
+  eprint = {1204.3007},
+  pages = {1321--1329},
+  publisher = {{International Union of Crystallography}},
+  issn = {00218898},
+  doi = {10.1107/S0021889812043026},
+  urldate = {2014-01-16},
[…]
