+++ paper.md
@@ -1,0 +1,61 @@
+---
+  title: 'PASCal Python: A Principal Axis Strain Calculator'
+  tags:
+  - Python
+  - crystallography
+  - non-ambient
+  - thermal expansion
+  - compressibility
+  - electrochemical strain
+
+  authors:
+    - name: Monthakan Lertkiattrakul
+      affiliation: 1
+    - name: Matthew L. Evans
+      orcid:  0000-0002-1182-9098
+      affiliation: 2
+    - name: Matthew J. Cliffe
+      orcid: 0000-0002-0408-7647
+      affiliation: 1
+  affiliations:
+  - name: School of Chemistry, University Park, Nottingham, NG7 2RD, United Kingdom
+    index: 1
+  - name: Institut de la Matière Condensée et des Nanosciences, Université catholique de Louvain, Chemin des Étoiles 8, Louvain-la-Neuve 1348, Belgium
+    index: 2
+  date: 28 May 2023
+
+  bibliography: paper.bib
+...
+
+# Summary
+
+The response of crystalline materials to external stimuli: whether temperature, pressure or electrochemical potential, is critical for both our understanding of materials and their use. This information can be readily obtained through in-situ diffraction experiments, however if the intrinsic anisotropy of crystals is not taken into account, the true behaviour of crystals can be overlooked. This is particularly true for anomalous mechanical properties of great topical interest, such as negative linear or area compressibility [@Cairns2015; @Hodgson2014], negative thermal expansion [@chenNegativeThermalExpansion2015] or strongly anisotropic electrochemical strain [@kondrakovAnisotropicLatticeStrain2017].
+
+We have developed PASCal, Principal Axis Strain Calculator, a widely used web tool (https://github.com/MJCliffe/PASCal) that implements the rapid calculation of principal strains and fitting to many common models for equations of state. It provides a simple web form user interface designed to be able to be used by all levels of experience. This new version of PASCal is written in Python using the standard scientific Python stack [@harrisArrayProgrammingNumPy2020; @virtanenSciPyFundamentalAlgorithms2020], is released open source under the MIT license, and significantly extends the feature set of the original closed-source Fortran, Perl and Gnuplot webtool [@Cliffe2012b]. Significant additional attention has been paid to testing, documentation, modularisation and reproducibility, enabling the main app functionality to now also be accessed directly through a Python API. The web app is deployed online at [https://www.pascalapp.co.uk](https://www.pascalapp.co.uk) with the associated source code and documentation available on GitHub at [MJCliffe/PASCal](https://github.com/MJCliffe/PASCal).
+
+
+# Statement of Need
+
+Characterising the anisotropic strain response of a crystalline material requires calculating not just the deformation along unit cell axes directions, which for low symmetry crystals have an arbitrary orientation relative to the mechanical properties, but also how changes in unit cell angles affect the principal (eigen-)strains and their orientations (eigenvectors). A great deal of additional information can be extracted from the stimulus-dependence of these principal strains, and fitting this dependence to equations-of-state can produce parameters of fundamental importance (e.g., compressibilities). Software capable of transforming experimental lattice parameters into the principal strains and the derived underlying thermodynamic properties (e.g. expansivities or compressibilities) is therefore of great utility to experimental scientists studying the response of crystals.
+
[…]
