### 2023-09-22 7f8a557
Function definitions improved
