### 2023-09-22 #5556 @MJCliffe
Hi @phibeck, we have now hopefully implemented all the changes suggested. We now have greatly improved the usability and robustness of the software I hope, thanks in large part to @ml-evs .

- the web app can now be installed as a python package using pip
- implemented extensive unit tests;
- documentation much more extensive and includes example python notebooks
- stable weblink [www.pascalapp.co.uk](www.pascalapp.co.uk])

Matt has been added as a corresponding author to reflect his significant contribution to the software as a result. The paper and the branch to review are the main 'Live' branch and the software should be ready for re-review.
