+++ paper.bib
@@ -190,7 +190,8 @@
   title = {Collaborative Data Science},
   author = {{Plotly Technologies Inc}},
   year = {2015},
-  address = {{Montr\'eal, QC (Canada)}}
+  address = {{Montr\'eal, QC (Canada)}},
+  url = {{https://plot.ly}}
 }
 % == BibTeX quality report for plotlytechnologiesincCollaborativeDataScience2015:
 % ? unused Url ("https://plot.ly")
