### 2023-09-23 cfbe91c
attempt to get typing-extensions installed
