+++ .github/workflows/deploy.yml
@@ -4,6 +4,7 @@ on:
   release:
     types:
      - published
+  workflow_dispatch:
 
 jobs:
 
