### 2023-09-23 249b856
Force Python 3.9 in CI
