+++ .github/workflows/ci.yml
@@ -16,10 +16,10 @@ jobs:
       with:
         submodules: true
 
-    - name: Set up Python 3.8
+    - name: Set up Python 3.9
       uses: actions/setup-python@v4
       with:
-        python-version: '3.8'
+        python-version: '3.9'
         cache: 'pip'
         cache-dependency-path: |
           requirements*.txt
