+++ .github/workflows/ci.yml
@@ -6,6 +6,12 @@ on:
     branches:
     - Live
 
+# Cancel running workflows when additional changes are pushed
+# https://docs.github.com/en/actions/using-workflows/workflow-syntax-for-github-actions#example-using-a-fallback-value
+concurrency:
+  group: ${{ github.head_ref || github.run_id }}
+  cancel-in-progress: true
+
 jobs:
 
   tests:
