### 2023-09-23 eaffa59
Add explicit typing-extensions dep for Python <= 3.9, then lock deps
