### 2023-09-26 #5556 @MarDiehl
@ml-evs @MJCliffe nice improvements.

There are only 2 small things:
1) when running the most recent version of mypy, I get a bunch of errors:
[mypy.log](https://github.com/openjournals/joss-reviews/files/12731262/mypy.log). Not relevant for the review and most of them might be difficult to fix. The only quick fix is the replacement of deprecated `np.product` with `np.prod`.
2) there is no link in the plotly reference
