### 2023-09-27 ae7e016
Fix pre-commit setup with poetry
