+++ .pre-commit-config.yaml
@@ -31,3 +31,10 @@ repos:
     hooks:
       - id: ruff
         args: [--fix]
+
+  - repo: https://github.com/pre-commit/mirrors-mypy
+    rev: v1.5.1
+    hooks:
+      - id: mypy
+        name: "MyPy"
+        additional_dependencies: ["types-all"]
