### 2023-09-27 #5556 @ml-evs
> @ml-evs @MJCliffe nice improvements.
>
> There are only 2 small things:
>
>     1. when running the most recent version of mypy, I get a bunch of errors:
>        [mypy.log](https://github.com/openjournals/joss-reviews/files/12731262/mypy.log). Not relevant for the review and most of them might be difficult to fix. The only quick fix is the replacement of deprecated `np.product` with `np.prod`.
>

Thanks @MarDiehl, I've just remedied these errors (as you say, the numpy one was the only signficant one -- although it seems `np.product` is still maintained as an undocumented alias!) I've added mypy to our CI runs.

>     2. there is no link in the plotly reference

This should have been fixed in a previous version, but probably before I asked the bot to rerender the PDF. I'll double check that now.
