### 2023-09-27 #5556 @MarDiehl
@ml-evs: I get a deprecation warning for `np.product` and fixed it recently in my code. That's why I was alert to it.
