### 2023-09-27 8b80d41
Minor mypy tweaks
