+++ paper.bib
@@ -191,7 +191,7 @@
   author = {{Plotly Technologies Inc}},
   year = {2015},
   address = {{Montr\'eal, QC (Canada)}},
-  url = {{https://plot.ly}}
+  url = {https://plot.ly}
 }
 % == BibTeX quality report for plotlytechnologiesincCollaborativeDataScience2015:
 % ? unused Url ("https://plot.ly")
