### 2023-10-09 MJCliffe/PASCal#15 @lbluque
JOSS review
Hi @ml-evs  and @MJCliffe

I'm adding my review comments for openjournals/joss-reviews#5556 here.

Overall, the latest updates to the source code and documentation look great. I found it very straightforward to use the web tool and the API directly.

Specific comments:

### General

- Can you clarify the contributions of the first author M.L.? I notice they do not show up in the commit history.

### Functionality

- I had no problem running the app locally or accessing it on https://www.pascalapp.co.uk; however, the link in the README https://pascalapp.co.uk is not working for me.

### Documentation

- The web tool link in the documentation is somewhat hard to make out. It might be helpful to have a link to the web app more prominently located.
- I noticed the CI tests use Python 3.9, does the software run in newer versions? I think you should mention supported Python versions; additionally, since your unit tests run quite quickly you can run tests using all supported Python versions.
- I suggest adding a search bar to the documentation. I personally find it extremely helpful to be able to quickly reference API.
- The contributing page is very informative, though I would add the following details for developers: how to install with additional testing requirements and the recommended way to run tests. Personally, I have never used poetry, so this would be helpful to have mentioned upfront.

### Paper

- I noticed you have a list of other useful software in the documentation. It would be useful to have these and any other related software included in the paper, along with a brief discussion of how PASCal compares.
