### 2023-10-09 MJCliffe/PASCal#15 @MJCliffe
Thanks! We'll get workin on the issues, which seem sensible and manageable.

Regarding the contribution of Monthakan (ML), she wrote the initial version offline as a summer project, which formed the basis of the first functional version uploaded to github. Her summer project has finished so she is no longer an active contributor.
