+++ pyproject.toml
@@ -12,7 +12,10 @@ license = "MIT"
 authors = [ "Matthew Cliffe <matthew.cliffe@nottingham.ac.uk>", ]
 classifiers = [
     "Programming Language :: Python :: 3",
-    "Programming Language :: Python :: 3.8",
+    "Programming Language :: Python :: 3.9",
+    "Programming Language :: Python :: 3.10",
+    "Programming Language :: Python :: 3.11",
+    "Programming Language :: Python :: 3.12",
     "Intended Audience :: Science/Research",
     "Topic :: Scientific/Engineering",
 ]
@@ -22,7 +25,7 @@ repository = "https://github.com/MJCliffe/PASCal"
 packages = [{ include = "PASCal", from = "src" }]
 
 [tool.poetry.dependencies]
-python = "^3.8"
+python = "^3.9"
 Flask = "*"
 Jinja2 = "*"
 numpy = "*"
