+++ .github/workflows/ci.yml
@@ -17,15 +17,22 @@ jobs:
   tests:
     runs-on: ubuntu-latest
 
+    strategy:
+      fail-fast: false
+      max-parallel: 4
+      matrix:
+        python-version: ['3.9', '3.10', '3.11', '3.12']
+
     steps:
     - uses: actions/checkout@v3
       with:
         submodules: true
 
-    - name: Set up Python 3.9
+
+    - name: Set up Python ${{ matrix.python-version }}
       uses: actions/setup-python@v4
       with:
-        python-version: '3.9'
+        python-version: ${{matrix.python-version }}
         cache: 'pip'
         cache-dependency-path: |
           requirements*.txt
@@ -37,6 +44,7 @@ jobs:
         poetry install --all-extras
 
     - name: Run linting
+      if: matrix.python-version == 3.9
       run: |
         poetry run pre-commit run --all-files
 
@@ -45,10 +53,12 @@ jobs:
         poetry run pytest
 
     - name: Build docs
+      if: matrix.python-version == 3.9
       run : |
         poetry run mkdocs build
 
     - name: Lock dependencies and prepare deployment
+      if: matrix.python-version == 3.9
       run: |
         # Azure expects to find the app, requirements and any static files at the top-level
         # so we need to do some tweaks here to massage the repo into that format
@@ -60,6 +70,7 @@ jobs:
         ls -al
 
     - name: Upload artifact for deployment
+      if: matrix.python-version == 3.9
       uses: actions/upload-artifact@v3
       with:
         name: PASCal - test
@@ -71,12 +82,14 @@ jobs:
           !tests/
 
     - name: Fix permissions # followng https://github.com/actions/upload-pages-artifact?tab=readme-ov-file#file-permissions
+      if: matrix.python-version == 3.9
       run: |
         chmod -c -R +rX "site/" | while read line; do
           echo "::warning title=Invalid file permissions automatically fixed::$line"
         done
 
     - name: Upload docs artifact
+      if: matrix.python-version == 3.9
       uses: actions/upload-pages-artifact@v2
       with:
         path: site
