### 2023-10-09 e458da2
Pin flask to v2 and add upper bound of Python 3.13 to allow installation of new numpy
