+++ mkdocs.yml
@@ -50,6 +50,8 @@ extra_javascript:
   - https://cdn.jsdelivr.net/npm/mathjax@3/es5/tex-mml-chtml.js
 
 plugins:
+  - search:
+      lang: en
   - mkdocstrings:
       default_handler: python
       handlers:
