### 2023-10-09 571529d
Add README badges for web app, JOSS and docs
