### 2023-10-09 35c542f
Disable py312 due to dep incompatibility
