### 2023-10-09 1afcf0d
Add dev installation instructions to contributing page
