### 2023-10-10 0497f80
Update paper.md
added description of other software
