### 2023-10-10 fce913c
Update paper.bib
added ELATE citation
