+++ paper.bib
@@ -76,6 +76,17 @@
 % == BibTeX quality report for Cliffe2012b:
 % ? unused Url ("http://scripts.iucr.org/cgi-bin/paper?S0021889812043026")
 
+@article{ELATE2016,
+  title={ELATE: an open-source online application for analysis and visualization of elastic tensors},
+  author={Gaillac, Romain and Pullumbi, Pluton and Coudert, Fran{\c{c}}ois-Xavier},
+  journal={Journal of Physics: Condensed Matter},
+  volume={28},
+  number={27},
+  pages={275201},
+  year={2016},
+  publisher={IOP Publishing}
+}
+
 @book{Giacovazzo2011,
   title = {Fundamentals of {{Crystallography}}},
   author = {Giacovazzo, C and Monaco, H. L. and Artioli, G. and Viterbo, D. and Milanesio, M. and Ferraris, G. and Gilli, G. and Gilli, P. and Zanotti, G. and Catti, M.},
