### 2023-10-10 2fe8996
Use raw HTML README badges for the sake of the docs
