+++ paper.md
@@ -31,7 +31,7 @@
 
 The response of crystalline materials to external stimuli: whether temperature, pressure or electrochemical potential, is critical for both our understanding of materials and their use. This information can be readily obtained through in-situ diffraction experiments, however if the intrinsic anisotropy of crystals is not taken into account, the true behaviour of crystals can be overlooked. This is particularly true for anomalous mechanical properties of great topical interest, such as negative linear or area compressibility [@Cairns2015; @Hodgson2014], negative thermal expansion [@chenNegativeThermalExpansion2015] or strongly anisotropic electrochemical strain [@kondrakovAnisotropicLatticeStrain2017].
 
-We have developed PASCal, Principal Axis Strain Calculator, a widely used web tool (https://github.com/MJCliffe/PASCal) that implements the rapid calculation of principal strains and fitting to many common models for equations of state. It provides a simple web form user interface designed to be able to be used by all levels of experience. This new version of PASCal is written in Python using the standard scientific Python stack [@harrisArrayProgrammingNumPy2020; @virtanenSciPyFundamentalAlgorithms2020], is released open source under the MIT license, and significantly extends the feature set of the original closed-source Fortran, Perl and Gnuplot webtool [@Cliffe2012b]. Significant additional attention has been paid to testing, documentation, modularisation and reproducibility, enabling the main app functionality to now also be accessed directly through a Python API. The web app is deployed online at [https://www.pascalapp.co.uk](https://www.pascalapp.co.uk) with the associated source code and documentation available on GitHub at [MJCliffe/PASCal](https://github.com/MJCliffe/PASCal).
+We have developed PASCal, Principal Axis Strain Calculator, a widely used web tool that implements the rapid calculation of principal strains and fitting to many common models for equations of state. It provides a simple web form user interface designed to be able to be used by all levels of experience. This new version of PASCal is written in Python using the standard scientific Python stack [@harrisArrayProgrammingNumPy2020; @virtanenSciPyFundamentalAlgorithms2020], is released open source under the MIT license, and significantly extends the feature set of the original closed-source Fortran, Perl and Gnuplot webtool [@Cliffe2012b]. Significant additional attention has been paid to testing, documentation, modularisation and reproducibility, enabling the main app functionality to now also be accessed directly through a Python API. The web app is deployed online at [https://www.pascalapp.co.uk](https://www.pascalapp.co.uk) with the associated source code and documentation available on GitHub at [MJCliffe/PASCal](https://github.com/MJCliffe/PASCal).
 
 
 # Statement of Need
