### 2023-10-12 #5556 @hungpham2017
@phibeck thank you for being patient with my slow progress. After the following comment is considered/addressed. I am happy to recommend its publication in JOSS.
@ml-evs: just one comment:  In the **Comparison with other software** section of the manuscript, while several packages with related features are mentioned, there isn't a clear discussion on how PASCal stands out or offers superior capabilities in certain areas compared to these tools. It would be beneficial to include brief comments highlighting the unique or enhanced features of PASCal. This would provide clarity to readers and save them the effort of deducing these differences themselves.
