### 2023-10-13 #5556 @ml-evs
> @ml-evs: just one comment: In the **Comparison with other software** section of the manuscript, while several packages with related features are mentioned, there isn't a clear discussion on how PASCal stands out or offers superior capabilities in certain areas compared to these tools. It would be beneficial to include brief comments highlighting the unique or enhanced features of PASCal. This would provide clarity to readers and save them the effort of deducing these differences themselves.

I have not used any of these packages personally so @MJCliffe is probably better placed to answer, but the manuscript states that they are all closed source and/or desktop GUI driven, or simply "related" tools that do not address the same challenge as PASCal.
