### 2023-10-13 #5556 @hungpham2017
> As the manuscript is quite concise and now includes descriptions of the features of both PASCal's and the other software suites, I am not sure that adding extensive comparisons feature by feature will make the paper more straightforward for potential users. If there was a particular point of uncertainty that did not come through, we'd be happy to clarify, but otherwise my preference is for a more focussed paper.
>
> To amplify @ml-evs comment: the manuscript does already state that the other software tools that implement some of the PASCal feature set are closed source; and highlights that the only other online tool can only calculate single strain points. I am slightly reluctant to write more extensively about other software suite's feature set, particularly one that is under active development and does not have a roadmap or online documentation (e.g. EoSFit has recently subsumed WinStrain and is likely to evolve further). My hope is that as PASCal is online and is designed to be easy to use, users can try it (or consult the docs) to see whether it suits their analysis needs without significant time investment.

This response is clear. I aimed to underscore that readers and users can quickly see this project's advantages as an open-source and online tool, a distinction not often seen in similar software in the Comparison section. The decision to highlight this further rests with the authors. @phibeck the manuscript is good to go to me. I have completed my portion of the review.
