### 2023-10-13 #5556 @phibeck
Thank you. I would have actually suggested mentioning this point directly in the repo's README (since this is what potential users are likely to see first), but I think the first paragraph in the README already addresses this point to some extent. Feel free to expand on this, otherwise the review process is complete! Thank you, @MarDiehl, @lbluque and @hungpham2017 for your thorough reviews!

@MJCliffe @ml-evs  the reviewers have recommended the submission for publication. There are a few more steps before we finalize the publication.
