### 2023-10-14 0c9f47e
Add new citation info
