### 2023-10-15 #5556 @MJCliffe
- Authors and affiliations checked.
- Release of the software v2.1.0 made.
- Zenodo archive of release: 10.5281/zenodo.10004384
- Authors and title correct on archive
- Licence matches
