### 2023-10-18 #5556 @phibeck
@MJCliffe, a few more things:

- could you comment on the contributions by Monthakan Lertkiattrakul, since I could not find any commits by this author in the history
- could you please add the missing DOI indicated above

Some more minor remarks on the manuscript:
- line 16: either include link with url or remove it, since it's referenced in line 26
- line 53: remove additional "the"
- line 71: I would find the following sentence easier to understand, if this is what you mean: "as crystallographic errors propagating through to the strain are not in general well-defined"
- line 113: please use subscripts in Ag3[Co(CN)6]
