+++ paper/paper.bib
@@ -84,6 +84,7 @@
   number={27},
   pages={275201},
   year={2016},
+  doi = {10.1088/0953-8984/28/27/275201},
   publisher={IOP Publishing}
 }
 
