### 2023-10-18 2305f1c
description of error propagation improved
