### 2023-10-18 #5556 @MJCliffe
Thank you for the proof reading, I have made the changes.

Re: line 71, I've rephrased to hopefully make this clearer:
"as the propagation of the errors in the unit cell parameters through to the errors in the diagonalised strain is not well-defined in general"

@ml-evs will need to check that I've not introduced any additional errors.

Monthakan wrote the first functional draft of the program offline, without using github, as a summer project. Once the summer project ended I revised the software and uploaded it to github.
@lbluque also raised this as an issue: https://github.com/MJCliffe/PASCal/issues/15#issuecomment-1753520477
