### 2023-10-18 #5556 @phibeck
@openjournals/dev can you help us understand why editorialbot won't generate the pdf? Thank you!
