### 2023-10-19 #5556 @xuanxu
> can you help us understand why editorialbot won't generate the pdf? Thank you!

There's [an error](https://github.com/openjournals/joss-papers/actions/runs/6560948635/job/17856626482#step:3:762) coming from the compilation task. Summoning @tarleb as our Pandoc expert.
