+++ paper/paper.bib
@@ -102,7 +102,7 @@
 % ? unused Number of pages ("842")
 
 @article{Goodwin2008a,
-  title = {Large Negative Linear Compressibility of {{Ag{{$_3$}}}}[{{Co}}({{CN}}){{$_6$}}].},
+  title = {Large Negative Linear Compressibility of Ag{$_3$}[Co(CN){$_6$}].},
   author = {Goodwin, Andrew L and Keen, David A and Tucker, Matthew G},
   year = {2008},
   month = dec,
