### 2023-10-19 #5556 @MJCliffe
The pandoc error wasn't super informative for me, but I've tried fixing the bib as the mostly likely error.
