### 2023-10-20 1db46c6
Use Better BibTeX's idea of the best formatting of reference
