### 2023-10-20 b8e4fe7
Use Better BibTeX's idea of the best formatting of reference (#24)
