### 2023-10-20 #5556 @tarleb
I tried to compile locally, which worked. Maybe the LaTeX compiler is running out of memory, but I don't understand why.
