### 2023-10-20 #5556 @ml-evs
So to summarize, it seems the build system is having issues with the way we format chemical formulae in the reference for https://doi.org/10.1073/pnas.0804789105, "Large negative linear compressibility of Ag<sub>3</sub>[Co(CN)<sub>6</sub>]". I've now reverted to the version exported by the Zotero Better BibTeX plugin (which uses \textsubscript instead of mathmode, and I can confirm this works locally with pandoc 2.18 + LaTeX... so slightly at a loss how to proceed.
