### 2023-10-20 #5556 @tarleb
Can we try with `{{ Ag\textsubscript{3}{[}Co(CN)\textsubscript{6}{]} }}` ? Maybe the blank `[` and `]` are confusing the LaTeX parser.
