### 2023-10-20 #5556 @ml-evs
> Can we try with `{{ Ag\textsubscript{3}{[}Co(CN)\textsubscript{6}{]} }}` ? Maybe the blank `[` and `]` are confusing the LaTeX parser.

No dice... I tried specifically `{{Ag}}{\textsubscript{3}}{{[}}{{Co}}({{CN}}){\textsubscript{6}}{{]}}`, i.e., encasing the square brackets with double braces. Again this builds fine in the previewer...
