### 2023-10-20 #5556 @ml-evs
Ok... one more test to see if it is literally just the `[` that are breaking it...
