### 2023-10-20 #5556 @tarleb
Oh, what's interesting is that I can reproduce the issue by copying the LaTeX source code from the logs and trying to compile locally.
