### 2023-10-20 #5556 @ml-evs
That's helpful -- I still can't pin down what is failing here after stripping out large chunks of the bib. I've now reverted back entirely to the last bib that seemed to work here, to make sure its not a regression. We will then need to find a way to re-apply the suggestions of the editor without breaking it!
