### 2023-10-20 #5556 @tarleb
I think the problem is the line

        \item \href{https://doi.org/\url{10.5281/zenodo.10004384}}{\color{linky}{Archive}} \ExternalLink

The nested `\url` command is causing issues.
