### 2023-10-20 #5556 @ml-evs
> Done! archive is now [[10.5281/zenodo.10004384](10.5281/zenodo.10004384)]([https://doi.org/[10.5281/zenodo.10004384](10.5281/zenodo.10004384)](https://doi.org/%5B10.5281/zenodo.10004384%5D(10.5281/zenodo.10004384)))

Probably just this command was incorrect? Need to say

"@editorialbot set 10.5281/zenodo.10004384 as archive" rather than "@editorialbot set https://doi.org/10.5281/zenodo.10004384 as archive"
