### 2023-10-20 #5556 @ml-evs
I've just reverted the paper to its initial state before all this debugging -- fairly confident it will build now once the DOI is set correctly. I guess @phibeck is the one that needs to re-run the command? Then hopefully we're good to go! Thanks @tarleb for your help, and well spotted!
