### 2023-10-20 #5556 @xuanxu
Yeah, nice detective work! The origin of that value for the archive DOI in the PDF metadata comes from the incorrect value in the top of the issue.
