### 2023-10-20 #5556 @xuanxu
This should fix the PDF compilation
