### 2023-10-20 #5556 @ml-evs
> MISSING DOIs
>
> - 10.1088/0953-8984/28/27/275201 may be a valid DOI for title: ELATE: an open-source online application for analysis and visualization of elastic tensors
>
> INVALID DOIs
>
> - None
> ```
>

Let me just double check this one!
