### 2023-10-20 #5556 @phibeck
> Just had to resync the branches, the DOI for ELATE should be there now.
>
> Could you trigger the final procedure again please @phibeck? (Also absolutely no problem about the DOI command!)

I'm not sure I need to trigger it again actually. I assume the AEiC @Kevin-Mattheus-Moerman was pinged already and will take over from here.
