### 2023-10-23 #5556 @ml-evs
Thanks everyone involved!

> [![DOI](https://camo.githubusercontent.com
/ab508ed34897a5016bc1b59e1af35534a5eb048dcf2ca87c37bde086f8c13a4d/68747470733a2f2f6a6f73732e7468656f6a2e6f72672f7061706572732f31302e32313130352f6a6f73732e30353535362f7374617475732e737667)](https://doi.org/10.21105/joss.05556)

Just a minor point/heads-up, do the DOI badges look okay for everyone else? On the latest Firefox Linux & Windows it looks like this:

![2023-10-23-180408_233x54_scrot](https://github.com/openjournals/joss-reviews/assets/7916000/f58d7cd1-ebaa-4f53-a5d1-5526c87bbc57)

This is also the case for the other GitHub "native" status badges for CI workflows... Not urgent to fix if but just wanted to point it out!
